% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrettyBarPlot2.R
\name{areColors}
\alias{areColors}
\title{Check that colores names are the names of some real colors.}
\usage{
areColors(x)
}
\arguments{
\item{x}{a string vectors with (to be checked) naesm of colors.}
}
\value{
x a logical vector with \code{TRUE} if a color, \code{FALSE} if not.
}
\description{
\code{areColors}: Check that colores names are the names of some real colors.
}
\details{
credit:
\url{Credit: https://stackoverflow.com/questions/13289009/check-if-character-string-is-a-valid-color-representation}
}
\examples{
areColors(c(NA, "black", "blackk", "1", "#00", "#000000"))
}
