% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSCA.R
\name{blockProj}
\alias{blockProj}
\title{Compute the partial projections for blocks in
3-way Correspondence Analysis (as in block-PTCA)}
\usage{
blockProj(data, b.weights, nI, FS, sv, data.metric = NULL)
}
\arguments{
\item{data}{An \eqn{I} * \eqn{J} data matrix structured in
\eqn{K} blocks of rows, each described by the same
\eqn{J} variables.}

\item{b.weights}{the weights for the \eqn{K} blocks.
Can be a \eqn{K} by 1 vector or a scalar if all blocks
have same weight.}

\item{nI}{the number of rows of the blocks,
Can be a \eqn{K} by 1 vector or a scalar if all blocks
have same number of rows.}

\item{FS}{An \eqn{I} * \eqn{L} matrix (with
\eqn{L}: number of factors of the analysis) storing
the factor scores for the \eqn{I}-set obtained from
 correspondence analysis.}

\item{sv}{The singular values obtained from the
analysis of the whole data table.}

\item{data.metric}{(a \eqn{J} by 1 vector) stores
the metric for the columns of \code{X}. When
\code{NULL} (default) \code{blockProj} uses
the standard transition formula and first transform the
data into colum profiles. When the CA was not a standard CA
(i.e., different centers or different metrics),
\code{data.metric} is used to normalize the data matrix
prior to the barycentric projection.}
}
\value{
a \eqn{J} (variables) by \eqn{L} (factors)
by \eqn{K} (blocks) array storing the \eqn{K} "slices"
pf partial factor scores.
}
\description{
\code{blockProj}:
Computes the partial projections for blocks in
3-way Correspondence Analysis (e.g., as in block-PTCA).
Compute the projection for blocks on one set from
the recontsitution formula from the other set.
}
\details{
This function is used when the orignal data table is
made of blocks or rows (resp columns) all described by the
same columns (resp rows). In the row case, the rows
are clustered in blocks of size
\eqn{I_}1, ... ,\eqn{I_k}, ... \eqn{I_K}
(with sum \eqn{I_K} = \eqn{I}) all described by \eqn{J} variables
(i.e., columns). \code{blockProj} computes the variables
(i.e., the \eqn{J}-set) partial projections
of the \eqn{K} blocks of rows. The partial projections
are barycentric to the whole set of projections
(i.e., the barycenters off all \eqn{K}-blocks is
equal to the factors for the whole matrix.
the projection is obtained from the standard reconstitution
formula generalized to the case of blocks with each block
having a \eqn{b}-weight
(with \eqn{b_k} > 0, and sum \eqn{b_k} = 1n).
When the data are obtained from a standard CA, the
paramter \code{data.metric} does not need to be used.
When the data are obtained from an unconventional CA
(i.e., as performed  with the decomposition
in common and specific factors).
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
genCA
}
\author{
Herve Abdi
}
