% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootRatio.R
\name{boot.ratio.test}
\alias{boot.ratio.test}
\title{boot.ratio.test computes bootstrap ratios
from a "bootstrap cube"}
\usage{
boot.ratio.test(boot.cube, critical.value = 2)
}
\arguments{
\item{boot.cube}{An I*L*B bootstrap brick
(typically obtained from Boot4PTCA).
The third dimension (B) corresponds to the
random factor of the
"to-be-bootstrapped-units"
(e.g., judges, participants, assessors)}

\item{critical.value}{The critical value for significance
(default = 2, which matches a \eqn{p <} .05 significance level)}
}
\value{
A list: 1) sig.boot.ratios: A logical vector that identifies the
significant items. 2)  boot.ratios: the bootstrap ratios
3) prob.boot.ratios: the (uncorrected) probability associated to the
bootstrap ratios 4) prob.boot.ratios.cor:
the (Sidak/Bonferonni) corrected probability associated to the
bootstrap ratios
}
\description{
boot.ratio.test computes bootstrap ratios
from a "bootstrap cube" (created, e.g., by
BOOT4PTCA)
}
\examples{
#  BR <- boot.ratio.test(BootI)
}
\author{
Derek Beaton & Hervé Abdi
}
