% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boot4Mean_etc.R
\name{Boot4Mean}
\alias{Boot4Mean}
\title{Boot4Mean computes a brick of bootstrap
estimates for the means of groups of observations
described by several variables (e.g., factor scores).}
\usage{
Boot4Mean(Data, design, niter = 100, suppressProgressBar = TRUE)
}
\arguments{
\item{Data}{An observations by variables data set.}

\item{design}{A vector giving the design for the observations:
observations with the same number belong to the same group}

\item{niter}{(Default = \code{100})
the number of bootstrap iterations.}

\item{suppressProgressBar}{(default = \code{TRUE}), when
\code{TRUE} suppress the progress bar.}
}
\value{
A list with \code{BootCube:} a groups by
variables by iterations brick of bootstrapped means, and
\code{GroupMeans}: the groups by variables
original means of the groups and
\code{BootsrappedGroupMeans}:
the groups by variables bootstrapped means of
\code{BootCube}.
}
\description{
\code{Boot4Mean} computes a brick of bootstrap
estimates for the means of groups of observations
described by several variables,  as for example,
for groups of observations described by factors scores
computed from a Principal Component Analysis or
or a Multiple Correspondence Analysis.
The groups of observations are identified by a design factor.
Observations are Bootstrapped within their groups
(so each Bootstrap sample has the  same
N per group as the original groups).
}
\details{
Use with PCA-like techniques as
a step before plotting confidence interval ellipses for
group means
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Herve Abdi
}
