% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootRatio.R
\name{Boot4PTCA}
\alias{Boot4PTCA}
\title{Boot4PTCA. Compute the Bootstrapped Factors scores
for the  I and J sets from
a Partial Triadic Correspondence analysis
(PTCA)}
\usage{
Boot4PTCA(ZeDataCube, fi, fj, eigs, nf2keep = 2, nBootIter = 100,
  compact = FALSE)
}
\arguments{
\item{ZeDataCube}{An I*J*K data cube (K are observations)
The third dimension (K) is bootstrapped}

\item{fi}{The factor scores for I (rows) from the epCA program}

\item{fj}{The factor scores for J (columns) from the epCA program}

\item{eigs}{The eigenvalues from the epCA program}

\item{nf2keep}{how many factors to keep,  default to 2}

\item{nBootIter}{How many Bootstrap samples, default to 100
(RowsBoot = ZeBootCube_I,ColumnsBoot = ZeBootCube_J)}

\item{compact}{(default = FALSE) if \code{TRUE} gives a compact
version with only the result for the symmetric approch.}
}
\value{
A list: if compact false:
1a) RowsBoot a I*L*B cube of Bootstrapped
coordinates for the I-set
1b) RowsBoot.asym a I*L*B cube of Bootstrapped
coordinates for the I-set
(asymmetric projections);
 2a)  ColumnsBoot a J*L*B cube of Bootstrapped coordinates
 for the J-set
 if compact false 2b)
 ColumnsBoot.asym a J*L*B cube of Bootstrapped
 coordinates for the J-set
 with I: number of rows, J: number of columns
 L: number of factors kept (i.e., nf2keep),
 B: number of Bootstrap replicates (i.e., nBootIter)
}
\description{
This function bootstraps the Kth dimension of a data cube
and computes bootstraped factor scores
}
\examples{
#  BootFactorsIJ <-Boot4PTCAt(A.Cube.Of.Data,fi=fi,fj=fj,eigs=eigs)
}
\author{
Herve Abdi
}
