% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick2Mat.R
\name{brick2Mat}
\alias{brick2Mat}
\title{Unfold a brick of data to a matrix}
\usage{
brick2Mat(aBrick, names4Rows = NULL)
}
\arguments{
\item{aBrick}{an \eqn{I} by \eqn{J} by \eqn{K}
brick of data.}

\item{names4Rows}{(Default: \code{NULL}) a \eqn{I}*\eqn{K}
vector for the row names of the output matrix.
When \code{NULL}, the row names are obtained by concatenation
of the \code{dimnames(brick2mat)[[1]]} and
\code{dimnames(brick2mat)[[3]]}.}
}
\value{
an
(\eqn{I}*\eqn{K}) by \eqn{J} matrix.
}
\description{
\code{brick2Mat}:
Unfold an \eqn{I} by \eqn{J} by \eqn{K}
brick of data to an
(\eqn{I}*\eqn{K}) by \eqn{J} matrix that will inherit
the dimension names from the brick of data.
}
\author{
Herve Abdi
}
