% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization4CA.R
\name{CARenormalization}
\alias{CARenormalization}
\title{re-normalize CA factor scores to Asymetric and
True-Barycentric factor scores}
\usage{
CARenormalization(G, delta, singularValues = TRUE, masses = NULL)
}
\arguments{
\item{G}{a set of factor scores from a correspondence analysis}

\item{delta}{a vector of singular / eigen values}

\item{singularValues}{do we have singular values or eigenvalues
when TRUE (default) we have singular values, if FALSE we have
eigenvalues}

\item{masses}{a vector of masses for tor the observations. if NULL
do not compute the Biplot normalization}
}
\value{
a list with G_A (Asymmetric factor scores), G_B
 (True Barycentric factor scores), G_S (SPSS Symmetric Biplot),
 and G_P (biPlot).
}
\description{
Re-normalize a set of CA factor scores
to be 1) Asymmetric: points are the vertices of the simplex,
(Inertia per dimension = Eigenvalues = 1);
2) True Barycentric: points of one set are at the
true barycenter of the points of the other set
(Inertia per dimension = Eigenvalues^2 of the
standard-aka symmetric analysis);
3) Biplots (see Greenacre, 2007, p. 102): The asymetric set
is rescaled by the inverse square root of its masses;
4) The strange SPSS renormalization
where the factor scores
have variance equal to the singular values.
These re-normalizations schemes make sense only when
the sets of items for rows and columns are asymmetric
 such as,for example, when one set represents an
 independent variable and the other one a dependent variable
 (e.g., "check all that apply" CATA data).
 When plotting both sets on the same map,
  the independent variable will be the set
 with the largest inertia.
 This type of plot is possible with two
 different configurations:
 1)  IV-set with Asymmetric and DV-set Symmetric, or
 2) IV-set with symmetric and DV-set with True Barycentric.
}
\author{
Herve Abdi
}
