% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanZeData.R
\name{cleanDataCATA}
\alias{cleanDataCATA}
\title{cleanDataCATA
Clean a Brick of Data prior to a CATA analysis}
\usage{
cleanDataCATA(listCATA, threshold4cleaning = NULL)
}
\arguments{
\item{listCATA}{a List as obtained, for example, from
\code{read.xls.CATA}.}

\item{threshold4cleaning}{(default = 10% of # judges) the minimum number
of choices for a variable to be kept. 
default is 10.}
}
\value{
A list with \code{$CATA.Brick}
a product by adjective by judge Brick of 0/1 data
(1 if judge chose adjective for product, 0 if not);
 \code{$CATA.CT}: 
  A product by adjective cleaned contingency table
  (columns whose sum is equal or below threshold are dropped).
  \code{$nProducts} the number of products,
  \code{$nJudges} the number of judges, and
  \code{$nVars} the number of variables kept.
}
\description{
cleanDataCATA cleans a CATA Brick 
  of data prior to a CATA analysis.
  The CATA Brick is
  obtained, for example, from \code{PTCA4CATA::read.xls.CATA}.
  This list should include $CATA.Brick.
  The level of "almost empty" is defined by the parameter
  \code{threshold4cleaning} 
  (set by default to 10% of the number of judges).
}
\author{
Herve Abdi
}
