% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFactorMaps4CA.R
\name{createAllMaps4CA}
\alias{createAllMaps4CA}
\title{Creates all the partial ggplot2 maps for CA
with all standard variants of normalization for factor scores.}
\usage{
createAllMaps4CA(allNormedFactors, axis1 = 1, axis2 = 2, title = NULL,
  col.points.i = "blueviolet", alpha.points.i = 0.5, pch.i = 19,
  cex.i = 2.5, segment.size.i = 0, col.labels.i = "darkorchid4",
  alpha.labels.i = 1, text.cex.i = 4, font.face.i = "bold",
  font.family.i = "sans", force.i = 1, col.points.j = "darkolivegreen4",
  alpha.points.j = 0.5, pch.j = 18, cex.j = 2.5, segment.size.j = 0,
  col.labels.j = "darkolivegreen", alpha.labels.j = 1, text.cex.j = 4,
  font.face.j = "bold", font.family.j = "sans", force.j = 1,
  col.axes = "darkorchid", alpha.axes = 0.2, width.axes = 1.1,
  col.background = adjustcolor("lavender", alpha.f = 0.2), ...)
}
\arguments{
\item{allNormedFactors, }{A list with all the
factor scores normed
typically from createAllNormedFactors}

\item{axis1}{the column of X used for the horizontal axis
of the plot. Default 1.}

\item{axis2}{the column of X used for the vertical axis
of the plot. Default 2.}

\item{title}{A main title (default is NULL).}

\item{col.points.i}{the color of the points/dots for the I-set.
Can be one color or a vector of colors. If a vector it needs
to have exactly the number of items to be plotted.
Default = 'blueviolet'.}

\item{alpha.points.i}{(default = .5), the alpha
(transparency) for the points, should be
between 1 (no transparency) and 0
(completely transparent).}

\item{pch.i}{the character for the points for the I-set.
Default is 19 (Circles).}

\item{cex.i}{size of the dots for the I-set. Default = 2.5}

\item{segment.size.i}{= 0, # size of segment}

\item{col.labels.i}{the color of the labels for
the I-set.
Can be one color or a vector of colors. If a vector, it needs
to have exactly the number of items to be plotted.
Default =  'darkorchid'.}

\item{alpha.labels.i}{(default = 1), the alpha
(transparency) for the labels, should be
between 1 (no transparency) and 0
(completely transparent).}

\item{text.cex.i}{= 4,  font size for labels for the I-set.}

\item{font.face.i}{=   'bold',  font for labels for the I-set.}

\item{font.family.i}{= 'sans' , font family for the I-set.}

\item{force.i}{= 1. How much ggrepel repels the labels
for the  I-set.
#' for the  I-set.}

\item{col.points.j}{the color of the points/dots for the J-set.
Can be one color or a vector of colors. If a vector it needs
to have exactly the number of items to be plotted.
Default = 'darkolivegreen4'.}

\item{alpha.points.j}{(default = .5), the alpha
(transparency) for the J-points, should be
between 1 (no transparency) and 0
(completely transparent).}

\item{pch.j}{the character for the points for the J-set.
Default is 18 (Diamonds).}

\item{cex.j}{size of the dots for the J-set. Default = 2.5}

\item{segment.size.j}{= 0, # size of segment for J-set}

\item{col.labels.j}{the color of the labels for
the J-set.
Can be one color or a vector of colors. If a vector, it needs
to have exactly the number of items to be plotted.
Default =  'darkolivegreen'.}

\item{alpha.labels.j}{(default = 1), the alpha
(transparency) for the J-labels, should be
between 1 (no transparency) and 0
(completely transparent).}

\item{text.cex.j}{= 4,  font size for labels for the J-set.}

\item{font.face.j}{=   'bold',  font for labels for the J-set.}

\item{font.family.j}{= 'sans' , font family for the J-set.}

\item{force.j}{= 1. How much ggrepel repels
the labels for the J-set.}

\item{col.axes}{= 'darkorchid', color for the axes}

\item{alpha.axes}{= .2 transoarency factors for the
color of the axes.}

\item{width.axes}{= 1.1,the width of the axes}

\item{col.background}{=  adjustcolor('lavender',
alpha.f = .2), The color of the background.}

\item{...}{eveythings else for the functions.}
}
\value{
a list
baseMap_S: The Background (Symmetric)
I_labels_S  The labels I-set.
I_points_S  The dots I-set.
J_labels_S  The labels J-set.
 J_points_S The dots J-set.
 baseMap_A The Background (Asymmetric)
  I_labels_A  The labels I-set.
 I_points_A The dots I-set.
  J_labels_A  The labels J-set.
  J_points_A  The dots J-set.
 baseMap_B The Background (Barycentric)
  I_labels_B The labels I-set.
  I_points_B  The dots I-set.
  J_labels_B The labels J-set.
 J_points_B  The dots J-set.
zeMap: The Complete Map (background Dots and Labels)
zeMap_dots:
zeMap_text:
factorScores:  The factor scores
constraints: The list of the contraints
}
\description{
\code{createAllMaps4CA } uses
 \code{ggplot2} to create all the partial maps for CA
with the different types of normalizations: Asymmetric,
Symmetric, True-Barycentric, Biplots and SPSS pseudo Biplots.
 Create ggplot2 factorial maps for dots and labels
create the base plot maps for CA type graphs with ggplot2.
Creates maps for the I (rows) and the J (column) -sets.
}
\seealso{
createFactorMaps createbaseMap
}
\author{
Herve Abdi
}
