% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization4CA.R
\name{createAllNormedFactors}
\alias{createAllNormedFactors}
\title{Create all the normalization for the factor scores of a CA:
Symmetric (the standard), Asymetric, and
True-Barycentric factor scores.}
\usage{
createAllNormedFactors(ResFrom.epCA, namesOfFactors = "Dimension")
}
\arguments{
\item{ResFrom.epCA}{The results from
\code{ExPosition::epCA}}

\item{namesOfFactors}{the names of the factors
(default is 'Dimension')}
}
\value{
a list with the normalized factor scores
 \code{Fi} (I-set original factor scores), \code{Fi_A}
 (I-set asymmetric normalization),
  \code{Fi_B} (I-set true barycentric),
  \code{Fi_P} (I-set Greenacre's Biplot),
  \code{Fi_S} (I-set SPSS strange biplot),
 \code{Fj} (J-set original factor scores),
 \code{Fj_A}
 (J-set asymmetric normalization),
  \code{Fj_B} (J-set true barycentric),
  \code{Fj_P} (J-set Greenacre's Biplot),
  \code{Fj_S} (J-set SPSS strange biplot)
  @examples
# Use the colorOfMusic data example
 data("colorOfMusic")
# NB need to have ExPosition installed
 resCA <- ExPosition::epCA(colorCT, graphs = FALSE)
 renormedFactors <- createAllNormedFactors(resCA)
}
\description{
Create all the different normalization for the rows
and columns of a correspondence analysis:
Re-normalize a set of CA factor scores
to be 1) Asymetric: points are 1) the vertices of the simplex,
(Inertia per dimension = Eigenvalues = 1) and
2) True Barycentric: points of one set are at the
true barycenter of the points of the other set
(Inertia per dimension = Eigenvalues^2 of the
standard-aka symmetric analysis).
These re-normalizations schemes make sense only when
the sets of items for rows and columns are asymmetric
 such as for example when one set represents an
 independent variable and the other one a dependent variable
 (e.g., "check all that apply" data).
 When plotting both sets on the same map,
  the independent variable will be the set
 with the largest inertia. This is possible with two
 different configurations:
 1)  IV-set with Asymmetric and DV-set Symmetric, or
 2) IV-set with symmetric and DV-set with True Barycentric.
}
\author{
Herve Abdi
}
