% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFactorMaps4CA.R
\name{createFactorMap}
\alias{createFactorMap}
\title{create the base plot maps for CA type graphs with ggplot2}
\usage{
createFactorMap(X, axis1 = 1, axis2 = 2, constraints = NULL,
  title = NULL, col.points = "blueviolet", alpha.points = 0.5,
  display.points = TRUE, pch = 19, cex = 2.5, display.labels = TRUE,
  col.labels = "darkorchid4", alpha.labels = 1, text.cex = 4,
  font.face = "bold", font.family = "sans", col.axes = "darkorchid",
  alpha.axes = 0.2, width.axes = 1.1,
  col.background = adjustcolor("lavender", alpha.f = 0.2), force = 1,
  segment.size = 0, ...)
}
\arguments{
\item{X}{the factor scores to plot}

\item{axis1}{the column of X used for the horizontal axis
of the plot. Default 1.}

\item{axis2}{the column of X used for the vertical axis
of the plot. Default 2.}

\item{constraints}{a list with minx miny maxx maxy
typically obtained from \code{ExPosition}. If NULL
(default) it  is computed with the function
 \code{prettyGraphs::minmaxHelper()}.}

\item{title}{A title for the graph. Default is NULL}

\item{col.points}{the color of the points/dots.
Can be one color or a vector of colors. If a vector it needs
to have exactly the number of items to be plotted.
Default = 'blueviolet'.}

\item{alpha.points}{(default = .5), the alpha
(transparency) for the points, should be
between 1 (no transparency) and 0
(completely transparent).}

\item{display.points}{if  TRUE (Default)
create the map for the points.}

\item{pch}{the character for the points,
Default is 19 (Circles).}

\item{cex}{size of the dots. Default = 2.5}

\item{display.labels}{if TRUE, create the
map for the labels. Default is TRUE.}

\item{col.labels}{the color of the labels
Can be one color or a vector of colors. If a vector it needs
to have exactly the number of items to be plotted.
Default =  'darkorchid'.}

\item{alpha.labels}{(default = 1), the alpha
(transparency) for the points, should be
between 1 (no transparency) and 0
(completely transparent).}

\item{text.cex}{= 4, font size for labels.}

\item{font.face}{= 'bold', font for labels.}

\item{font.family}{= 'sans',
font family for labels}

\item{col.axes}{color for the axes, default is
'darkorchid'}

\item{alpha.axes}{alpha parametere (transparency)
for the axes, default is .2}

\item{width.axes}{the width of the axes, default is 1.1}

\item{col.background}{the color theme of the background,
default is adjustcolor('lavender', alpha.f = .2)}

\item{force}{= 1. How much ggrepel repels the label}

\item{segment.size}{= 0,
size of segment line for ggrpel}

\item{...}{stuff to be passed to other functions.}
}
\value{
a list
zeMap: The Complete Map (background Dots and Labels)
zeMap_dots: The dots
zeMap_text: The Labels
factorScores:  The factor scores
constraints: The list of the contraints
}
\description{
Create the ggplot2 basic factor maps for CA type graphs.
The final maps are created by using overlays on top of the base map.
See also \code{makeCAmap}.
A map for CA is created with the baseMap to which
is added text/dot map.
\code{createFactorMap} calls the functions \code{map4DotsAndLabels}
and \code{ createBaseMap}.
}
\seealso{
\code{map4DotsAndLabels}
\code{ createBaseMap}.
}
\author{
Herve Abdi
}
