% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFactorMaps4CA.R
\name{createFactorMapIJ}
\alias{createFactorMapIJ}
\title{Create \code{createFactorMapIJ}
ggplot2 factorial maps for CA-type pf maps dots and labels.}
\usage{
createFactorMapIJ(Fi, Fj, axis1 = 1, axis2 = 2, constraints = NULL,
  title = NULL, col.points.i = "blueviolet", alpha.points.i = 0.5,
  pch.i = 19, cex.i = 2.5, segment.size.i = 0,
  col.labels.i = "darkorchid4", alpha.labels.i = 1, text.cex.i = 4,
  font.face.i = "bold", font.family.i = "sans", force.i = 1,
  col.points.j = "darkolivegreen4", alpha.points.j = 0.5, pch.j = 18,
  cex.j = 2.5, segment.size.j = 0, col.labels.j = "darkolivegreen",
  alpha.labels.j = 0.5, text.cex.j = 4, font.face.j = "bold",
  font.family.j = "sans", force.j = 1, col.axes = "darkorchid",
  alpha.axes = 0.2, width.axes = 1.1,
  col.background = adjustcolor("lavender", alpha.f = 0.2), nudge_x = 0,
  nudge_y = 0, ...)
}
\arguments{
\item{Fi}{the I-set factor scores to plot}

\item{Fj}{the J-set factor scores to plot}

\item{axis1}{the column of X used for the horizontal axis
of the plot. Default 1.}

\item{axis2}{the column of X used for the vertical axis
of the plot. Default 2.}

\item{constraints}{a list with minx miny maxx maxy
typically obtained from  \code{prettyGraphs::minmaxHelper()}.
If NULL
(default) it  is computed with the function
 \code{prettyGraphs::minmaxHelper()}.}

\item{title}{mina title for the plots}

\item{col.points.i}{the color of the points/dots for the I-set.
Can be one color or a vector of colors. If a vector it needs
to have exactly the number of items to be plotted.
Default = 'blueviolet'.}

\item{alpha.points.i}{(default = .5), the alpha
(transparency) for the points, should be
between 1 (no transparency) and 0
(completely transparent).}

\item{pch.i}{the character for the points for the I-set.
Default is 19 (Circles).}

\item{cex.i}{size of the dots for the I-set. Default = 2.5}

\item{segment.size.i}{= 0, # size of segment}

\item{col.labels.i}{the color of the labels for
the I-set.
Can be one color or a vector of colors. If a vector, it needs
to have exactly the number of items to be plotted.
Default =  'darkorchid'.}

\item{alpha.labels.i}{(default = 1), the alpha
(transparency) for the i-labels, should be
between 1 (no transparency) and 0
(completely transparent).}

\item{text.cex.i}{= 4,  font size for labels for the I-set.}

\item{font.face.i}{=   'bold',  font for labels for the I-set.}

\item{font.family.i}{= 'sans' , font family for the I-set.}

\item{force.i}{= 1. How much ggrepel repels the labels
for the  I-set.
#' for the  I-set.}

\item{col.points.j}{the color of the points/dots for the J-set.
Can be one color or a vector of colors. If a vector it needs
to have exactly the number of items to be plotted.
Default = 'darkolivegreen4'.}

\item{alpha.points.j}{(default = .5), the alpha
(transparency) for the points, should be
between 1 (no transparency) and 0
(completely transparent).}

\item{pch.j}{the character for the points for the J-set.
Default is 18 (Diamonds).}

\item{cex.j}{size of the dots for the J-set. Default = 2.5}

\item{segment.size.j}{= 0, # size of segment for J-set}

\item{col.labels.j}{the color of the labels for
the J-set.
Can be one color or a vector of colors. If a vector, it needs
to have exactly the number of items to be plotted.
Default =  'darkolivegreen'.}

\item{alpha.labels.j}{(default = 1), the alpha
(transparency) for the j-labels, should be
between 1 (no transparency) and 0
(completely transparent).}

\item{text.cex.j}{= 4,  font size for labels for the J-set.}

\item{font.face.j}{=   'bold',  font for labels for the J-set.}

\item{font.family.j}{= 'sans' , font family for the J-set.}

\item{force.j}{= 1. How much ggrepel repels
the labels for the J-set.}

\item{col.axes}{= 'darkorchid', color for the axes}

\item{alpha.axes}{= .2 transoarency factors for the
color of the axes.}

\item{width.axes}{= 1.1,the width of the axes}

\item{col.background}{=  adjustcolor('lavender',
alpha.f = .2), The color of the background.}

\item{nudge_x}{= 0. From ggrepel,
nudge value for starting point for
 labels: x dimension.}

\item{nudge_y}{= 0. From ggrepel,
nudge value for starting point for
 labels: y dimension.}

\item{...}{everything else for the functions}
}
\value{
a list
zeMap: The Complete Map (background Dots and Labels)
zeMap_dots: The dots
zeMap_text: The Labels
factorScores:  The factor scores
constraints: The list of the contraints
}
\description{
Create ggplot2 factorial maps for dots and labels.
NB needs a base map to work correctly.
#' create the base plot maps for CA type graphs with ggplot2
}
\seealso{
createFactorMaps createbaseMap
}
\author{
Herve Abdi
}
