% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot4CATA.R
\name{createLabel.gen}
\alias{createLabel.gen}
\title{Creates \eqn{x} and \eqn{y} labels for \code{ggplots2} scatterplots}
\usage{
createLabel.gen(zeAxis, lambda, tau, axisName = "Dimension ")
}
\arguments{
\item{zeAxis}{the number of the axis (no default)}

\item{lambda}{the eigen-value
associated with this dimension}

\item{tau}{the percentage of variance
associated with this dimension}

\item{axisName}{the name for the axes
(default = \code{'Dimension'})}
}
\description{
\code{createLabel.gen}
creates \eqn{x} and \eqn{y} labels for \code{ggplots2} scatterplots
(e.g., correspondence analysis or principal component analysis).
Compared to \code{createLabels},   \code{createLabels.gen}
does not require the results from \code{ExPosition}
and so can be used for any \code{ggplot2}-based scatterplot.
}
\author{
Herve Abdi
}
