% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMap4PartialFS.R
\name{createPartialFactorScoresMap}
\alias{createPartialFactorScoresMap}
\title{Create a \code{ggplot2} map for partial factor scores
(for \code{PTCA4CATA}, or\code{distatis}) to be added to
the main map of the factor scores.}
\usage{
createPartialFactorScoresMap(factorScores, partialFactorScores, axis1 = 1,
  axis2 = 2, colors4Items = NULL, colors4Block = NULL,
  names4Partial = NULL, alpha.lines = 0.5, size.lines = 0.75,
  type.lines = 1, arrow.length = 0, alpha.points = 0.7,
  shape.points = 23, size.points = 1, alpha.labels = 0.7,
  font.labels = "plain", family.labels = "sans", size.labels = 2)
}
\arguments{
\item{factorScores}{an \eqn{I} Items by \eqn{J}
factors/dimensions
data set storing the main factor scores.}

\item{partialFactorScores}{an
\eqn{I} Items by \eqn{J} factors/dimensions by
\eqn{K} Blocks storing the partial factor scores.
Needs to have numbers of Items as \code{factorScores}.}

\item{axis1}{(Default: 1) the horizontal dimension of the map.}

\item{axis2}{(Default: 2) the vertical dimension of the map.}

\item{colors4Items}{(Default: \code{NULL})
color name(s) for the Items,
can be one element or a \eqn{I} by 1 vector of color names.
If \code{NULL} (defaults), \code{prettyGraphsColorSelection()}
is used to select the colors.}

\item{colors4Block}{(Default: \code{NULL})
color name(s) for the Blocks,
can be one element or a \eqn{K} by 1 vector of color names.
If \code{NULL} (defaults), \code{prettyGraphsColorSelection()}
is used to select the colors.}

\item{names4Partial}{a vector for (preferably short)
names of the Blocks. If
\code{NULL} (default), the names are obtained from
the third dimension of \code{partialFactorScores};
if the third dimension of \code{partialFactorScores}
is \code{NULL}, then block are names from 1 to \eqn{K}.}

\item{alpha.lines}{(Default: 0.5) transparency parameter
for the lines, should be between 0 (complety transparent)
and 1 (no transparent).}

\item{size.lines}{(Default: 0.75) thickness of the lines
connecting the partial factor scores to their main scores.}

\item{type.lines}{(Default: 1) the type (from 1 to 6)
 of the lines
connecting the partial factor scores to their main scores.}

\item{arrow.length}{(Default: 0) the length in cm. of the
arrow point toward the factor scores.}

\item{alpha.points}{(Default: 0.7)
transparency parameter
for the points, should be between 0 (complety transparent)
and 1 (no transparent).}

\item{shape.points}{(Default: 23) the pch for the shape
(from 1 to 25). Default (23) is the diamond.}

\item{size.points}{(Default: 1) the size of the points.}

\item{alpha.labels}{(Default: 0.7)
transparency parameter
for the labels, should be between 0 (complety transparent)
and 1 (no transparent).}

\item{font.labels}{(Default: 'plain') the font for the
labels, can be \code{"plain", "bold", "italic"}}

\item{family.labels}{(Default: 'sans') the family of the
fontn used, this set depends upon the set of font installed,
the three save choices (always installed) are
\code{'sans','serif','mono'}.}

\item{size.labels}{(Default: 2) the size of the labels.}
}
\value{
a list with
1)  \code{$mapColByItems}:  the Partial Factor
Score Map Colored by Items;
2)
\code{$mapColByBlocks}: the Partial Factor
Score Map Colored by Blocks;
3) \code{$linesColByItems}: the Lines Colored by Items;
4) \code{$pointsColByItems}: Points Colored by Items;
5) \code{$labelsColByItems}: the Labels Colored by Items;
6) \code{$linesColByBlocks}:  the Lines Colored by Blocks;
7) \code{$pointsColByBlocks}: the Points Colored by Blocks;
and 8) #'\code{$labelsColByBlocks}:  the Labels Colored by Blocks.
}
\description{
\code{createPartialFactorScoresMap}:
Creates a map for partial factor scores for blocks of data
(e.g., for \code{PTCA4CATA} or \code{distatis}) to be added to
the main map of the factor scores (created for example
by \code{createFactorMap}).
The partial factor scores are colored by Blocks and by Items
(i.e. observations).
The data are 1) an \eqn{I} Items by \eqn{J} factors/dimensions
data set storing the main factor scores
(\code{factorScores}), and  2)
an \eqn{I} Items by \eqn{J} factors/dimensions by
\eqn{K} Blocks storing the partial factor scores.
The resulting map connects the partial factor scores to their
respective main factor scores.
}
\details{
The maps show the connections from the
partial factor scores to the main factor scores
(which are baycentric to their respective partial factor scores).
The map obtained should be added to a main map (of the
main factor scores) that should be created prior to
adding the map of the partial factor scores.
Because the partial factor scores have a larger variance
than the main factor scores, the dimension (i.e.
the \code{constraints}) of the
map should be computed (with \code{minmaxHelper4Partial()})
before creating the main of the main factor scores
(see example).
}
\examples{
\dontrun{
# fS are factor scores and pFS are partial factor scores
# first get the constraints
constraint4PFS <- minmaxHelper4Partial(fS, pFS)
# Get the background map
BaseMap.FS <- createFactorMap(X = fS ,
                              constraints = constraint4PFS)
# Create the map of the partial factor scores
map4pFS <- createPartialFactorScoresMap(fS, pFS)
map_FS_pFS <- BaseMap.FS + map4pFS
# To print the Map:
print(map_FS_pFS )
}
}
\seealso{
minmaxHelper4Partial createFactorMap DistatisR
}
\author{
Herve Abdi
}
