% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSCA.R
\name{CSCA}
\alias{CSCA}
\title{Common and Specific Correspondence Analysis
(CSDA)
of a set of \eqn{K} \eqn{I*J} matched matrices}
\usage{
CSCA(brickOfMat, nfact = 3, b = NULL)
}
\arguments{
\item{brickOfMat}{a \eqn{I} items by \eqn{J} descriptors *
\eqn{K} blocks (e.g., matrices) suitable
for correspondence analysis (i.e., all non-negative elements).}

\item{nfact}{(Default = 3) number of factors to keep.}

\item{b}{(Default = \code{NULL})
a \eqn{K} elements weight vector for the matrices
(should all be positive and sum to 1). When \code{NULL}
\code{CSCA} computes them as the sum of each matrix
divided by the grand total. Note that it is general better
to have pre-normalized the matrices
(e.g., with \code{normBrick4PTCA}),
so that all matrices have the same weight.}
}
\value{
A list with
1) \code{allMatrices.resCA} results for the analysis of
the whole set of matrices stacked on top of each other,
2) \code{sumOfMatrices.resCA}
results (from \code{ExPosition::epCA}) for the analysis of
the sum (i.e., average with CA) of all matrices,
3) \code{diffOfMatrices},
results for the analysis of
the difference of the matrices to their average (from 2),
4) \code{partialProjOnSum},
The projection as supplementary elements of the matrices
onto their average,
and 5) \code{RvCoefficients}
the matrix of \code{Rv}-coefficient between the matrices.
\code{allMatrices.resCA} and \code{ExPosition::epCA} are lists
containing a) \code{}
\code{fi}: the \eqn{I*K} by \code{nfact} matrix of the
row factor scores, b) \code{fj}:
  the \eqn{I*}\code{nfact}{*K} array
   by \code{nfact} array of the
column factor scores, c) \code{Dv}: the singular values,
d) \code{eigs}: the eigenvalues,
e) \code{tau}: the percentage of Inertia, and
f) \code{Inertia} the total inertia;
\code{partialProjOnSum} is a
list
containing a)
\code{fi}: the \eqn{I*K} by \code{nfact} matrix of the
(supplementary) row factor scores, b) \code{fj}:
  the \eqn{I*}\code{nfact}{*K} array
   by \code{nfact} array of the (supplementary)
column factor scores.
}
\description{
\code{CSDA}: implements
Common and Specific Correspondence Analysis
of a set of \eqn{K} \eqn{I*J} matched matrices.
}
\details{
The analysis of the three matrices whose results are
given in
\code{allMatrices.resCA}, \code{sumOfMatrices.resCA}, and
\code{diffOfMatrices} implement an ANOVA like decomposition
of the Inertia such that All = Sum + Difference or
\strong{X}_\eqn{K}  =  \strong{G} +
(\strong{X}_\eqn{K} - \strong{G}), with
\strong{X}_\eqn{K} being the set of matrices,
\strong{G} being the sum matrix
(which in CA is the barycenter of all the matrices)
and (\strong{X}_\eqn{K} - \strong{G})
being the set of the differences of all the matrices
to their barycenter.
The matrix \strong{X}_\eqn{K}
is obtained by stacking all the original matrices
on top of each other
(so \strong{X}_\eqn{K} is an \eqn{I*K} by \eqn{J} matrix);
the
matrix (\strong{X}_\eqn{K} - \strong{G})
is obtained by subtracting from each
matrix in \strong{X}_\eqn{K} the matrix \strong{G}
(so \strong{X}_\eqn{K} - \strong{G}
is an \eqn{I*K} by \eqn{J} matrix);
the matrix \strong{G}
is obtained as the sum of all the matrices
in \strong{X}_\eqn{K} (so \strong{G}
is a \eqn{I} by \eqn{J} matrix).
The analysis of the matrix \strong{G} is made with
a standard CA program, but the correspondence analysis
of matrices (\strong{X}_\eqn{K} - \strong{G})
needs a special
CA program because this CA uses the row and column
metrics from \strong{G},
\strong{X}_\eqn{K} uses the same centers as \strong{G},
whereas the (\strong{X}_\eqn{K} - \strong{G})
matrix is uncentered
(these analyses are performed
by the function \code{genCA}
that allow specific metrics and centers).
For the analysis of \strong{X}_\eqn{K} and
(\strong{X}_\eqn{K} - \strong{G})
the row factor scores (\code{fi}) are
computed from the plain \code{genCA}
analysis and the column factor
scores (\code{fj})
are obtained from partial projection using the
correspondence analysis transition formula adapted to blocks
of matrices. Note that in a two table version
the partial column factor scores will be identical
(and identical to the overall column factor score) and
so in this case, the overal column factor scores can
be plotted.
Note that the two table version
of  CSCA could be obtained
from the analysis of the
[\strong{X}_1 \strong{X}_2 || \strong{X}_2 \strong{X}_1]
circulant matrix
(see Greenacre, 2003).
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\references{
The ideas used here are derived from
1) Escofier, B., & Drouet, D. (1983). Analyse des différences
entre plusieurs tableaux de fréquences.
\emph{Les Cahiers de l'Analyse des Données, 8}, 491-499;
2) Greenacre, M. (2003). Singular value decomposition of
matched matrices. \emph{Journal of Applied Statistics, 30},
1101-1113; and 3)
Takane Y. (2014). \emph{Constrained Principal Component Analysis
and Related Techniques}, Boca Raton: CRC Press.
}
\seealso{
normBrick4PTCA genPCA

\code{\link[ExPosition]{epCA}}
}
