% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootRatio.R
\name{DataCheckMark2Cube}
\alias{DataCheckMark2Cube}
\title{DataCheckMark2Cube Create a cube of Data
from the results from "Check-Mark" CATA data.}
\usage{
DataCheckMark2Cube(DataChecks, NameOfDescriptor = NULL)
}
\arguments{
\item{DataChecks}{A Stimuli by Participants table of checks
The Stimuli are the columns of DataChecks,
 The Participants are the rows of DataChecks,
The Descriptors are the numbers in DataChecks
    (i.e., 5 for X[2,3] means that Participant 2,
        choosed Descriptor 5 for Stimulus 3)}

\item{NameOfDescriptor}{a length K vector of names of
the descriptors. if NULL (default) descriptors are}
}
\value{
a Stimuli*Descriptors*Participants brick of counts
}
\description{
Create a cube of Data from the results of a "Check-Mark"
data set collected in DataChecks
These data correspond to participants matching one descriptor to
each stimulus of a set of stimuli
The Stimuli are the columns of DataChecks,
 The Participants are the rows of DataChecks,
The Descriptors are the numbers in DataChecks
    (i.e., 5 for X[2,3] means that Participant 2,
        choosed Descriptor 5 for Stimulus 3)
}
\examples{
# le.grey <- vec2gray(1:10)
}
\author{
Hervé Abdi
}
