% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perm4PTCA.R
\name{eig4CA}
\alias{eig4CA}
\title{computes the eigenvalues for
correspondence analysis (CA)}
\usage{
eig4CA(X)
}
\arguments{
\item{X}{a matrix of non-negative numbers}
}
\description{
eig4CA computes the eigenvalues for
correspondence analysis (CA). Needs a matrix
of non-negative numbers. The trivial eigenvalue
equals to 1, is not computed.
NB for computational efficiency,
 the matrix diagonalized is NOT the
matrix diagonalized in CA but it has the
same eigenvalues.
NB:  rows and columns with 0 sums have an inverse set to 0
to avoid infinities.
}
\author{
Herve Abdi
# @export
}
