% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSCA.R
\name{genCA}
\alias{genCA}
\title{Generalized Correspondence Analysis with choice of
row and column metrics and choice of row and column centers}
\usage{
genCA(X, nfact = 3, normalize.X = TRUE, r.metric = NULL,
  c.metric = NULL, r.center = NULL, c.center = NULL)
}
\arguments{
\item{X}{an \eqn{I} by \eqn{J} data table suitable for
correspondence analysis.}

\item{nfact}{(Default: 3) number of factors to keep.}

\item{normalize.X}{(Default: \code{TRUE}), when
\code{TRUE} normalize \code{X} to a sum of 1 (as in standard CA).}

\item{r.metric}{(Default: \code{NULL}),
an eqn{I} by 1 vector of the row-metric.
when \code{NULL}
use the standard row-metric from CA
i.e., the  inverse of the sum of the rows of
normalized X).}

\item{c.metric}{(Default: \code{NULL}),
a \eqn{J} by 1 vector of the columns-metric,
when \code{NULL}
use the standard column-metric from CA
i.e., the  inverse of the sum of the columns of
normalized X).}

\item{r.center}{(Default: \code{NULL}),
an \eqn{I} by 1 vector of the row-center.
when \code{NULL}
use the standard row-center from CA
i.e., the sum of the rows of
normalized X).}

\item{c.center}{(Default: \code{NULL}),
a \eqn{J} by 1 vector of the row-center.
when \code{NULL}
use the standard column-center from CA
i.e., the sum of the colums of
normalized X).}
}
\value{
A list with 1) $\code{fi}: the row factor scores, 2)
2) $\code{fj}: the columns factor scores.
}
\description{
\code{genCA}: Implements
Generalized Correspondence Analysis with choice of
row and column metrics and choice of row and column centers.
}
\details{
The current version is not optimized for
computational efficiency and uses the
general SVD function \code{ExPosition::genPDQ()}.
}
\examples{
# Use the colorOfMusic Example from PTCA4CATA
data(colorOfMusic)
resCA <- genCA(colorOfMusic$contingencyTable)
# gives a standard CA of these data
}
\author{
Herve Abdi
}
