% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeans.R
\name{getMeans}
\alias{getMeans}
\title{get statistics (e.g., \code{means})
 of groups of observations.}
\usage{
getMeans(G, factor, FUN = mean)
}
\arguments{
\item{G}{a data frame observations * variables.}

\item{factor}{a factor vector for the groups of observations}

\item{FUN}{(\code{default = mean}) the statistics to be computed.}
}
\value{
a data frame with the group means as rows and the
variables as columns.
}
\description{
\code{getMeans}: compute statistics
(e.g., \code{means}) by groups of observations stored in an
observations * variables
data-frame or matrix.
}
\examples{
toto = matrix(round(runif(12)*10), nrow = 6, ncol = 2)
getMeans(toto, factor(c(1,1,1,2,2,2)), median) # median by groups
}
\seealso{
\code{\link{MakeCIEllipses}}
}
\author{
Hervé Abdi
}
