% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTIEllipses.R
\name{ggConvexHull}
\alias{ggConvexHull}
\title{ggConvexHull use ggplot2 to plot
a peeled convex hull for a set of points
described by 2 variables.}
\usage{
ggConvexHull(data, x_axis = 1, y_axis = 2, percentage = 0.66,
  col.line = "darkorchid4", alpha.line = 1, line.size = 1,
  line.type = 1, col.hull = "darkorchid", alpha.hull = 0.4,
  names.of.factors = "Dimension ")
}
\arguments{
\item{data}{an item by variables data,frame or matrix.}

\item{x_axis}{column number for the first variable.
\code{Default = 1}}

\item{y_axis}{column number for the second variable.
\code{Default = 2}}

\item{percentage}{proportion of the
convexHull to keep, Default: .66}

\item{col.line}{= the color of the line for the hill.
Default is 'darkorchid4'.}

\item{alpha.line}{= 1,
alpha value (transparency) for the lines,
takes values between 0 (completely transparent) and
1 (no transparent).}

\item{line.size}{= 1,
thickness of the line for the ellipses.}

\item{line.type}{= 1 (solid line)
the type of line (values from 0 to 6).}

\item{col.hull}{=  'darkorchid',
the color to fill the hull.}

\item{alpha.hull}{= .4
alpha value (transparency) for the hull.}

\item{names.of.factors}{= "Dimension",
names of the factors.
 if NULL   name.of.factors is dimnames(data)[2],
 if dimnames(data)[2] is NULL, the dimensions
 will be labelled "Dimension".
 This parameter is
 needed to avoid conflict when plotting
 these names as they need to be the same as the names of the
 data used to make the BaseMap plot (i.e., Fi/Fj/Fij).}
}
\value{
a ggplot2 object
 to draw a convex hull
}
\description{
\code{ggConvexHull}
computes a peeled convex hull for a set of points
described by 2 variables and plot it.
Uses function \code{PeelZeHull}
derived from
the function \code{ExPosition::peeledHull} which, in turn,
got inspiration from Michael Greenacre package \code{CA}.
The original code can be found at
\url{http://carme-n.org/?sec=code2}.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Herve Abdi
}
