% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdrawPolygon.R
\name{ggdrawPolygon}
\alias{ggdrawPolygon}
\title{Draw a Polygon in a \code{ggplot2} based graph}
\usage{
ggdrawPolygon(F, order2draw = 1:nrow(F), color = "darkorchid",
  linetype = 3, size = 1, fill = "darkorchid", alpha = 0.6, ...)
}
\arguments{
\item{F}{a set of 2-dimensional coordinates}

\item{order2draw}{(Default: \code{1:nrow(F)}) the order of the
points to draw the polygon.}

\item{color}{(Default: 'darkorchid') the color of the line
of the polygon.}

\item{linetype}{(Default: \code{3}, i.e., dotted line), 
the line type to draw the polygon.}

\item{size}{(Default: \code{1}) the size of the line.}

\item{fill}{(Default: \code{'darkorchid'}) the color to fill-in
the polygon.}

\item{alpha}{(Default: \code{.4}) the transparency fa ctor
for the fill-in color of the polygon (gors from 0 completely
transparent to 1: opaque).}

\item{...}{other stuff to be passed to \code{ggplot2::geom_polygon}
if needed.}
}
\value{
a polygon to be added to a \code{ggplots2} image.
}
\description{
\code{ggdrawPolygon}: Draws a Polygon in a 
\code{ggplot2}-based factorial graph (e.g., from CA or PCA).
}
\details{
\code{ggdrawPolygon} is mostly a wraper
around \code{ggplot2::geom_polygon}.
}
\seealso{
\code{\link[ggplot2]{geom_polygon}}
}
\author{
Hervé Abdi
}
