% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot4CATA.R
\name{GraphPTCABoot}
\alias{GraphPTCABoot}
\title{GraphPTCABoot plots the bootstrapped results of a CA}
\usage{
GraphPTCABoot(
  FS,
  FBoot,
  axis1 = 1,
  axis2 = 2,
  item.colors = NULL,
  ZeTitle = "PTCA-Bootstrap",
  constraints = NULL,
  nude = FALSE,
  Ctr = NULL,
  lwd = 3.5,
  ellipses = TRUE,
  fill = TRUE,
  fill.alpha = 0.27,
  percentage = 0.95,
  dev.new = TRUE,
  new.plot = TRUE,
  cex = 1,
  text.cex = 0.8,
  pos = 2
)
}
\arguments{
\item{FS}{(factor scores) The coordinates
of the items to be plotted.}

\item{FBoot}{The Bootstrap cube of factor scores.}

\item{axis1}{(default is 1)  position in \code{FS}
of the  horizontal axis}

\item{axis2}{(default is 2),  position in \code{FS}
of the  vertical axis}

\item{item.colors}{(default is \code{NULL}),
a color matrix/vector for the items
 (could be provided from a previous call to the graph routine).}

\item{ZeTitle}{(default is  'PTCA-Bootstrap') a Title for the plot}

\item{constraints}{(default is \code{NULL})
the dimensions of the plot as a list
giving the constraints
= list(minx=number1,miny=number2, maxx=number3,maxy=number4)
often provided from a previous plot.}

\item{nude}{(default is \code{FALSE}),
when \code{TRUE} do not plot names.}

\item{Ctr}{(default is \code{NULL}),
if \code{TRUE}  have the dots for the item
 proportional to their contributions to the plane.}

\item{lwd}{(default is  3.5)  line width for the ellipse
(from dataEllipse).}

\item{ellipses}{(default is \code{TRUE})
when \code{TRUE} plot ellipses,
when \code{FALSE} plot convex hulls,}

\item{fill}{(default is \code{TRUE})
when \code{TRUE}
fill in the ellipses with lighter colors.}

\item{fill.alpha}{(default is .27)  alpha-transparency for filling
(see parameter \code{fill}) ellipses/convex-hulls.}

\item{percentage}{(default is 0.95)  critical value for the CIs}

\item{dev.new}{(default is \code{TRUE})  when \code{TRUE}
create a new window for plotting.}

\item{new.plot}{when  \code{TRUE} (default)  Clean the devise}

\item{cex}{(default is 1)   size of the dots}

\item{text.cex}{(default is  .8) # size of text
(to avoid conflicts with cex)}

\item{pos}{(default is 2)  position of the text
can be a scalar or a vector}
}
\value{
PTCA.out: the output from prettyPlot
}
\description{
\code{GraphPTCABoot} plots the results of a CA
with boostrap computed  Ellipsoids
or Convex Hulls confidence interval
needs packages \code{prettyGraphs} (for the graph)
and the convex Hulls
and \code{car} (for the ellipses).
}
\author{
Hervé Abdi
}
