% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootRatio.R
\name{InertiaTable}
\alias{InertiaTable}
\title{A function to compute the inertia of a contingency table}
\usage{
InertiaTable(X)
}
\arguments{
\item{X}{A contingency table (non negative numbers)}
}
\value{
the inertia (\emph{a la} correspondence analysis)
of the contingency table
}
\description{
\code{InertiaTable}:  computes the inertia
(i.e., phi square or chi-square / N)
of a contingency table. The computation matches the value obtained
from the inertia of the correspondence analysis
of the contingency table.
\code{InertiaTable} is mostly used for permutation tests
and bootstrap estimates.
}
\examples{
set.seed(42)
X = matrix(round(runif(12)*20), nrow = 3)
InertiaOfATable <- InertiaTable(X)
}
\author{
Hervé Abdi
}
