% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSCA.R
\name{ldiag}
\alias{ldiag}
\title{Left (i.e., pre) Multiply a matrix by a diagonal matrix}
\usage{
ldiag(y, X)
}
\arguments{
\item{y}{a \eqn{I} element
vector (of the diagonal elements of an \eqn{I} by \eqn{I} matrix)}

\item{X}{an \eqn{I} by \eqn{J} matrix.}
}
\value{
an \eqn{I} by \eqn{J} matrix equal
to diag(\strong{y}) %*% \strong{X}.
}
\description{
\code{ldiag}: Left (i.e., pre) Multiply
a matrix by a diagonal matrix (with only
 the diagonal elements being given).
}
\examples{
\dontrun{
 Z = ldiag(y, X)
 }
}
\seealso{
\code{\link{rdiag}}
}
\author{
Hervé Abdi
}
