% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrettyBarPlot2.R
\name{lighten}
\alias{lighten}
\title{gives a lighter shade of a color}
\usage{
lighten(colors, factor = 1.4, maxColorValue = 255)
}
\arguments{
\item{colors}{a vector of color names.}

\item{factor}{(Default: 1.4) lightening factor}

\item{maxColorValue}{(Default: 255) the maximum possible for this}
}
\value{
A lightened color
}
\description{
\code{lighten} gives a lighter shade of a color.
}
\details{
Credit to Roland, see:
\url{https://stackoverflow.com/questions/30219738/is-there-a-way-to-programmatically-darken-the-color-given-rgb-values}
}
\examples{
lighten(c('#FF0000FC','#E6E6FAFC'), 1.3)
}
