% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggEllipses.R
\name{MakeCIEllipses}
\alias{MakeCIEllipses}
\title{\code{MakeCIEllipses}.
Add Confidence interval ellipses to a CA-like plots.}
\usage{
MakeCIEllipses(data, axis1 = 1, axis2 = 2, names.of.factors = "Dimension",
  col = NULL, centers = NULL, line.size = 1, line.type = 1,
  alpha.ellipse = 0.3, alpha.line = 0.5, p.level = 0.95)
}
\arguments{
\item{data}{A cube of bootstraped factor scores
should be an I Items * K factors * nIter array
(e.g. as given by \code{epCA.inference.battery})
with: I number of items,
      K number of factors (at least 2)
      nIter number of Bootstrap samples}

\item{axis1}{= 1 the horizontal axis}

\item{axis2}{= 2 the vertical axise}

\item{names.of.factors}{= 'Dimension',
names of the factors.
 if NULL   name.of.factors is dimnames(data)[2],
 Needed to avoid conflict when plotting
 these names need to be the same as the names of the
 data used to make the BaseMap plot (i.e., Fi/Fj/Fij)}

\item{col}{= NULL
a string or a vector of strings with color names
if NULL use prettyGraphs scheme with
\code{prettyGraphs::prettyGraphsColorSelection}}

\item{centers}{= NULL
if NULL centers the ellipses on their barycenter
to center on another canter (i.e., item factor scores)
provide an I * K data frame or matrix}

\item{line.size}{= 1
thickness of the line for the ellipses}

\item{line.type}{= 1]
the type of line ofr the ellipses}

\item{alpha.ellipse}{= .3
alpha value (transparency) for the ellipses}

\item{alpha.line}{= .5,
alpha value (transparency) for the lines}

\item{p.level}{= .95
p value for the CI}
}
\value{
LeGraph.elli
}
\description{
MakeCIEllipses. Create the Confidence Intervals (CI) ellipses
for plots for the I or J sets of a CA or STATIS type of analysis
The results of \code{MakeCIEllipses} should be added to BaseMap
created, for example, by the function \code{CreateBaseMap()}
}
\examples{
\dontrun{
Fij = rbind(Fi,Fj) # Fi and Fj being factor scores say from ExPosition
ABaseMap     <- CreateBaseMap(Fij)
# Boot_I is a Bootstrap cube for the I-set
GraphElli <- MakeCIEllipses(Boot_I)
ABaseMap + GraphElli # plot the I Set CI ellipses
}
}
\author{
Herve Abdi
}
