% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatMap4CT.R
\name{makeHeatMap4CT}
\alias{makeHeatMap4CT}
\title{makeHeatMap4CT makes a ggplot2 HeatMap representation of
a contingency table (i.e. CATA data)}
\usage{
makeHeatMap4CT(aContingencyTable, colorAttributes = NULL,
  colorProducts = NULL, angle.x = 70, fontSize.x = 10, fontSize.y = 20)
}
\arguments{
\item{aContingencyTable}{a contingency table to be plotted}

\item{colorAttributes}{color for the names of the columns.
Can be one value or a vector. If NULL (default), colors
are chosen using \code{prettyGraphs::prettyGraphsColorSelection()}}

\item{colorProducts}{color for the names of the rows.
Can be one value or a vector. if NULL use dark grey.}

\item{angle.x}{(defalt = 70)
The slope when writing the name of the variables}

\item{fontSize.x}{(default = 10), font size for the name
of the variables}

\item{fontSize.y}{(default = 20), font size for the name
of the products}
}
\value{
a ggplot2 heatmap
}
\description{
\code{makeHeatMap4CT} makes a ggplot2 HeatMap representation of
a contingency table (e.g., CATA data).
makeHeatMap4CT assumes the contingency table
contains only non-negative numbers.
}
\examples{
\dontrun{
aHeatMap <- makeHeatMap4CT(someContingencyTable)
# where \\code{someContingencyTable} is a contingency table
}
}
\author{
Herve Abdi
}
