% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTIEllipses.R
\name{MakeToleranceIntervals}
\alias{MakeToleranceIntervals}
\title{\code{MakeToleranceIntervals}.
Add Tolerance interval hulls or
ellipses to factor score plots
(e.g., PCA, CA).}
\usage{
MakeToleranceIntervals(data, design, axis1 = 1, axis2 = 2,
  names.of.factors = "Dimension", col = NULL, centers = NULL,
  line.size = 1, line.type = 1, alpha.ellipse = 0.3, alpha.line = 0.5,
  p.level = 0.95, type = "hull")
}
\arguments{
\item{data}{A set of I*L factor scores
(as obtained, for example, from \code{ExPosition::epCA})}

\item{design}{an I by 1 factor giving the group membership
of the Ith observations.}

\item{axis1}{= 1 the horizontal axis}

\item{axis2}{= 2 the vertical axise}

\item{names.of.factors}{= 'Dimension',
names of the factors.
 if NULL   name.of.factors is dimnames(data)[2],
 if dimnames(data)[2] is NULL, the dimensions
 will be labelled "Dimension". This paramter is needed
 to avoid conflict when plotting
 these names as they must be the same as the names of the
 data used to make the BaseMap plot (i.e., Fi/Fj/Fij)}

\item{col}{= NULL
a string or a vector of strings with color names
if NULL use prettyGraphs scheme with
\code{prettyGraphs::prettyGraphsColorSelection}}

\item{centers}{= NULL
if NULL centers the ellipses on their barycenter
to center on another center (i.e., item factor scores)
provide an I * K data frame or matrix.}

\item{line.size}{= 1
thickness of the line for the ellipses}

\item{line.type}{= 1
the type of line for the ellipses}

\item{alpha.ellipse}{= .3
alpha value (transparency) for the ellipses.}

\item{alpha.line}{= .5,
alpha value (transparency) for the lines.}

\item{p.level}{= .66
p value for the TI}

\item{type}{(Default = \code{'hull'})
type of interval can be c('ellipse','hull')}
}
\value{
LeGraph.elli a graph with convex hulls
or ellipse to be added to the base map as created
by the function  \code{CreateBaseMap()}.
}
\description{
\code{MakeToleranceIntervals}:
 Creates the ToleranceIntervals (CI) convex hulls or ellipses
for plots for the I or J sets of a CA, PCA,  MFA,
STATIS, etc. type of analysis
The results (Hull or Ellipses)
of \code{MakeToleranceIntervals}
should be added
to the BaseMap
created (for example) by the function
 \code{CreateBaseMap()}.
}
\examples{
\dontrun{
Fij = rbind(Fi,Fj)
# Fi and Fj being factor scores say from ExPosition
ABaseMap <- CreateBaseMap(Fij)
# Here design.matrix is design matrix describing K groups in Fi
GraphHull <- MakeToleranceIntervals(Fi,design.matrix)
print(ABaseMap + GraphHull)  # plot the K TI ellipses
}
}
\author{
Herve Abdi
}
