% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perm4PTCA.R
\name{Malinvaud4ptca}
\alias{Malinvaud4ptca}
\title{computes the Malinvaud-Saporta test for
significance in correspondence analysis
                  with asymptotically
                  (and, if provided permutation)
                   derived p-values}
\usage{
Malinvaud4ptca(Data, res.perm4ptca, ndigit4print = 4)
}
\arguments{
\item{Data}{the contingency table used for CA}

\item{res.perm4ptca}{The results from perm4PTCA
that gives the eigenvalues obtained from a permutation test}

\item{ndigit4print}{(Default = 4) number of digits
to use to print the results}
}
\value{
: A dataframe with the results of the test
}
\description{
\code{Malinvaud4ptca}: Compute the Malinvaud-Saporta
test for significance
in Correspondence Analysis. Provides asymptotic
Chis-square based p-values, and if a set
of permuted eigenvalues is provided
(obtained, e.g., from \code{Perm4MultCA} or
from \code{perm4PTCA})
provides _p_-values based on permutation tests.
References:
Malinvaud, E. (1987). Data Analysis in applied socio-economic statistics
with special considerations of correspondence analysis.
Marketing Science Conference Proceedings, HEC-ISA, Jouy-en-Josas.
Also cited in Saporta (2011).
Probabilité et Analyse des Données (3rd Ed).
   Technip, Paris. p. 209.
}
\author{
Herve Abdi
}
