% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFactorMaps4CA.R
\name{map4DotsAndLabels}
\alias{map4DotsAndLabels}
\title{Create ggplot2 factorial maps for dots and labels.
NB needs a base map to work correctly}
\usage{
map4DotsAndLabels(data, axis1 = 1, axis2 = 2, display.points = TRUE,
  col.points = "blueviolet", alpha.points = 0.5, pch = 19, cex = 2.5,
  display.labels = TRUE, col.labels = "darkorchid4", alpha.labels = 1,
  force = 1, segment.size = 0, text.cex = 4, font.face = "bold",
  font.family = "sans", nudge_x = 0, nudge_y = 0, ...)
}
\arguments{
\item{data}{the factor scores to plot}

\item{axis1}{the column of X used for the horizontal axis
of the plot. Default 1.}

\item{axis2}{the column of X used for the vertical axis
of the plot. Default 2.}

\item{display.points}{if  TRUE (Default)
create the map for the points.}

\item{col.points}{the color of the points/dots.
Can be one color or a vector of colors. If a vector it needs
to have exactly the number of items to be plotted.
Default = 'blueviolet',}

\item{alpha.points}{(default = .5), the alpha
(transparency) for the points, should be
between 1 (no transparency) and 0
(completely transparent).}

\item{pch}{the character for the points,
Default is 19 (Circles).}

\item{cex}{size of the dots. Default = 2.5}

\item{display.labels}{if TRUE, create the
map for the labels. Default is TRUE.}

\item{col.labels}{the color of the labels
Can be one color or a vector of colors. If a vector it needs
to have exactly the number of items to be plotted.
Default =  'darkorchid'.}

\item{alpha.labels}{(default = 1), the alpha
(transparency) for the labels, should be
between 1 (no transparency) and 0
(completely transparent).}

\item{force}{= 1. How much ggrepel repels the labels.}

\item{segment.size}{= 0,
size of segment line for ggrpel}

\item{text.cex}{= 4, font size for labels.}

\item{font.face}{= 'bold', font for labels.}

\item{font.family}{= 'sans',
font family for labels.}

\item{nudge_x}{= 0. From ggrepel,
nudge value for starting point for
 labels: x dimension.}

\item{nudge_y}{= 0. From ggrepel,
nudge value for starting point for
 labels: y dimension.
 @param alpha.point (default = .5), the alpha
 (transparency) for the points, should be
 between 1 (no transparency) and 0
 (completely transparent).}

\item{...}{eveythings else for the functions}
}
\value{
a list
zeMap: The Complete Map (background Dots and Labels)
zeMap_dots: The dots
zeMap_text: The Labels
factorScores:  The factor scores
constraints: The list of the contraints
}
\description{
Create ggplot2 factorial map for dots and labels.
NB needs a base map to work correctly.
#' create the base plot maps for CA type graphs with ggplot2
}
\seealso{
createFactorMaps createbaseMap
}
\author{
Herve Abdi
}
