% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick2Mat.R
\name{mat2Brick}
\alias{mat2Brick}
\title{tranform a matrix into a Brick of Data}
\usage{
mat2Brick(aMat, nBlocks, names4Rows = NULL, names4Blocks = NULL)
}
\arguments{
\item{aMat}{an (\eqn{I}*\eqn{K}) by \eqn{J}
data matrix.}

\item{nBlocks}{number of blocks (i.e., 3rd dimension of
the output array). No default.}

\item{names4Rows}{(Default: \code{NULL}) a \eqn{I}
element
vector for the row names of the output array.
When \code{NULL}, the row names are
created as {I}-1 to  I-\eqn{I}.}

\item{names4Blocks}{(Default: \code{NULL}) a \eqn{K}
element
vector for the names of  3rd-dimension of the output array.
When \code{NULL}, the row names are
created  as \eqn{B}-1 to \eqn{B}-K.}
}
\value{
an \eqn{I} by \eqn{J} by \eqn{K}
brick of data (i.e., array).
}
\description{
\code{mat2Brick}:
Unfold an (\eqn{I}*\eqn{K}) by \eqn{J} data matrix into
an \eqn{I} by \eqn{J} by \eqn{K}
brick of data (i.e., array).
}
