% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmaxHelper4Brick.R
\name{minmaxHelper4Brick}
\alias{minmaxHelper4Brick}
\title{Computes the x- and y-axes constraints
for all \code{prettyGraphs}  \code{PTCA4CATA} plotting functions
that display bootstrapped based confidence, tolerance, or
prediction intervals}
\usage{
minmaxHelper4Brick(aBrick, axis1 = 1, axis2 = 2, expandFactor = 1,
  trimPercent = 0, ...)
}
\arguments{
\item{aBrick}{a brick of bootstrapped factor scores
(e.g., from \code{Boot4PTCA}).}

\item{axis1}{the number of the factor for the x-axis (default = 1)}

\item{axis2}{the number of the factor for the y-axis (default = 2)}

\item{expandFactor}{expansion factor, if larger than 1
this is an expansion, equal to 1 (default) this does nothing,
smaller
than 1 this is a shrinkage.}

\item{trimPercent}{(default = 0) the proportion of extreme
values (small and large) to eiliminate.}

\item{...}{other arguments that could be passed to
\code{prettyGraphs::minmaxHelper}.}
}
\value{
minMaxList	A list with the following values:
\code{minx, miny, maxx, maxy}.
}
\description{
\code{minmaxHelper4Brick} Computes the x- and y-axes constraints
for all \code{prettyGraphs} and \code{PTCA4CATA} plotting functions
that displays bootstrapped based confidence, tolerance or
prediction intervals.  \code{minmaxHelper4Brick} transforms the
brick into a matrix and calls  \code{prettyGraphs::minmaxHelper}.
 \code{minmaxHelper4Brick} is typically used to get
 the dimensions for  \code{PTCA4CATA} graphics that use
 \code{MakeCIEllipses} or for  \code{DistatisR}) graphs such as
 \code{DistatisR::GraphDistatisBoot}.
}
\author{
Herve Abdi and Derek Beaton
}
