% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMap4PartialFS.R
\name{minmaxHelper4Partial}
\alias{minmaxHelper4Partial}
\title{Computes the x- and y- axis constraints for
graphs combininig a dataframe and a cube of partial factor sscores.
Used
for all \code{prettyGraphs} and \code{prettyGraph}-like
(e.d. \code{PTCA4CATA}) functions.}
\usage{
minmaxHelper4Partial(FactorScores, partialFactorScores, axis1 = 1,
  axis2 = 2, expansionFactor = 1.1)
}
\arguments{
\item{FactorScores}{An \eqn{I} observations by \eqn{J}
dimensions  data frame od factor scores}

\item{partialFactorScores}{An \eqn{I} observations by \eqn{J}
dimensions by \eqn{K} block array of partial factor scores.}

\item{axis1}{(Default: 1) the number of the first dimension
of interest.}

\item{axis2}{(Default: 2) the number of the second dimension
of interest.}

\item{expansionFactor}{(Default: 1.1) a multiplicative factor
for the list (when larger than 1 this is an expansion,
smaller than 1 this is a shrinkage).}
}
\value{
A list with \code{minx, miny, maxx, maxy} that can be
used by various graphic functions from, for example,
\code{prettyGraphs}, \code{DistatisR}, or \code{PTCA4CATA}.
}
\description{
\code{minmaxHelper4Partial}:
Computes the x- and y- axis constraints for
graphs combininig a dataframe and a cube of partial factor sscores.
Used
for all \code{prettyGraphs} and \code{prettyGraph}-like
(e.d. \code{PTCA4CATA}) functions. Factors scores and
partial factor scores should be describing the same set of
\eqn{I} observations.
}
\examples{
\dontrun{
testminmax <- minmaxHelper4Partial(someScores, somePartialScores )
}
}
