% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrangeJuiceCATARawData.R
\name{OrangeJuiceCATARawData}
\alias{OrangeJuiceCATARawData}
\title{\code{OrangeJuiceCATARawData} example of an excel file
with CATA data. Can be read by
\code{read.xls.CATA}.}
\description{
#' \code{OrangeJuiceCATARawData} example of an excel file
with CATA data. Can be read by
\code{read.xls.CATA}.
In this example of a "Check All That Apply" (CATA) task,
42 assessors evaluated 10 orange juices described by
33 descriptors.
To read this data use \code{system.file()}
(see example below).
}
\section{OrangeJuiceCATARawData.xlsx}{
NA
}

\examples{
path2file <- system.file("extdata", "OrangeJuiceCATARawData.xlsx", package = 'PTCA4CATA')
OrangeData <- read.xls.CATA(path2file = path2file, sheet2read = 'CATA')
}
\author{
Herve Abdi
}
