% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTstared.R
\name{p2Star}
\alias{p2Star}
\title{p2Star Converts a vector of probabilities into a vector
of stars for display.}
\usage{
p2Star(pValues, threshold = c(0.05, 0.01, 0.001))
}
\arguments{
\item{pValues}{a vector of $p$ values (i.e., values between 0 and 1)}

\item{threshold}{a vector giving the values for thresholding
default is c(.05,.01,.001)}
}
\value{
pStared a vector of stars
}
\description{
p2Star Converts a vector of probabilities into a string vector
of stars for display.
}
\examples{
\dontrun{
stars4p <- p2Star(aVectorOfPValues)
}

}
\author{
Herve Abdi
}
