% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockProjections.R
\name{partialProj4CA}
\alias{partialProj4CA}
\title{\code{partialProj4CA} Compute blocks (of variables)
partial projections for a Correspondence Analysis.}
\usage{
partialProj4CA(resCA, code4Blocks, rowBlocks = FALSE)
}
\arguments{
\item{resCA}{the results of the (CA) analysis from epCA,
for example \code{reFromCA <- epCA(X)}.}

\item{code4Blocks}{a vector indicating which variabless belong
to what block  (i.e.,
the variables of the same block have the same level for
\code{code4Block}): Needs to be of length equal to the
number of variables of the analysis.}

\item{rowBlocks}{= \code{FALSE} (default).
When \code{TRUE}, \code{partialProj4CA} runs the analysis
on blocks of observations instead of blocks of variables and
exchange the roles of the observations and variables.}
}
\value{
a list with (1) \code{Fk}:
an I*L*K array of the partial projections
of the K blocks, for the I observations, and the L factors
from \code{epCA}; (2) \code{Ctrk} an I*L matrix of the "relative"
block contributions [for a given component
the relative contributions sum to 1];
(3) \code{absCtrk} an I*L matrix of the "absolute"
block contributions [for a given component
the absolute contributions sum to the eigenvalue
for this component];
(4) \code{bk} a K*1 vector storing the weights for the blocks,
(5) \code{resRV} a list with
(a) a matrix storing the RV coefficients
between the blocks and, if the package \code{FactoMineR} is
installed, (b) the p-value for the RV-coefficient (as
computed with \code{FactoMineR::coeffRV}).
}
\description{
\code{partialProj4CA} Compute blocks (of variables)
partial projections for a Correspondence Analysis (CA).
Blocks are non-overlapping sets of variables (columns)
of a data table analyzed with
CA (as performed with \code{epCA} from \code{ExPosition}).
\code{partialProj4CA} gives
the partial projection for the blocks.
These projections are barycentric
because the barycenters of the
partial projections are equal to the factor scores
for the whole table.
}
\author{
Herve Abdi
}
