% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTIEllipses.R
\name{peelZeHull}
\alias{peelZeHull}
\title{Create a peeled convex hull for a set of points
described by 2 variables.}
\usage{
peelZeHull(data_matrix, x_axis = 1, y_axis = 2, percentage = 0.66)
}
\arguments{
\item{data_matrix}{an item by variables data,frame or matrix.}

\item{x_axis}{column number for the first variable.
\code{Default = 1}}

\item{y_axis}{column number for the second variable.
\code{Default = 2}}

\item{percentage}{proportion of the
convexHull to keep, Default: .66}
}
\value{
a dataframe with
 the coordinates of the peeled convex hull.
}
\description{
\code{peelZeHull}
computes a peeled convex hull for a set of points
described by 2 variables. Function taken from
the function \code{ExPosition::peeledHull} which, in turn,
got its inspiration from Michael Greenacre's package \code{CA}.
The original code can be found at
\url{http://carme-n.org/?sec=code2}.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\references{
for a reference
for the code for peeling the
convex hull see:  Greenacre, M. J. (2007).
Correspondence Analysis in Practice. Chapman and Hall.
}
\author{
Herve Abdi
}
