% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perm4PTCA.R
\name{perm4PTCA}
\alias{perm4PTCA}
\title{\code{perm4PTCA} Permutation test for the inertia and eigenvalues for PTCA}
\usage{
perm4PTCA(
  aCube,
  nIter = 1000,
  permType = "byRows",
  compact = FALSE,
  Malinvaud = TRUE
)
}
\arguments{
\item{aCube}{an I*J*K (i.e., products*Descriptors*Assessors)
of CATA Data.}

\item{nIter}{(Default = 1000). Number of Iterations
(i.e., number of permuted samples computed).}

\item{permType}{what type of permutation is used
if 'byRows' (default) the data are permuted within
the rows (i.e., products) for each assessor.
if 'byCols'  the data are permuted within
the columns (i.e., descriptors) for each assessor.
if 'byMat' the data are permuted within
the whole data matrix (i.e., products*descriptors)
for each assessor.}

\item{compact}{if TRUE return
(Default) only p-values for omnibus test}

\item{Malinvaud}{(default = TRUE) return the results
of the Malinvaud-Saporta chi square test}
}
\value{
a list with
\code{fixedInertia}: the CA-inertia of the data matrix;
\code{fixedEigenvalues}: the CA-eigenvalues of
the data matrix;
\code{pOmnibus}: the probability associated
to the inertia.
If \code{compact} is \code{FALSE}, return also
\code{permInertia}:
an \code{nIter} * 1 vector containing the
permutated inertia;
\code{pEigenvalues}: The probabilites
associated to each eigenvalue;
If \code{compact} is is \code{FALSE}, return also
\code{permEigenvalues}: an
\code{nIter} * \code{L} matrix giving
the permuted eigenvalues.
}
\description{
\code{perm4PTCA} Permutation for PTCA4CATA.
Computes an omnibus permutation test and
specific tests for the eigenvalues when
performing a PTCA analysis on Check All That Apply (CATA)
Data. Input is a Cube of data (products*Descriptors*Assessors).
Three different permutation schemes
are currently available (see paramater
\code{permType}).
}
\author{
Herve Abdi
}
