% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot4CATA.R
\name{PlotScree}
\alias{PlotScree}
\title{plot the scree for the eigen values
of an SVD based multivariate analysis}
\usage{
PlotScree(ev, p.ev = NULL, max.ev = NULL, alpha = 0.05,
  col.ns = "#006D2C", col.sig = "#54278F",
  title = "Explained Variance per Dimension")
}
\arguments{
\item{ev}{the eigen values to plot. no default}

\item{p.ev}{the probabilities associated to the
eigen-values}

\item{max.ev}{the max eigen-value
needed because ExPosition does not return all ev
but only the requested one. but return all tau
so if max.ev is specified, it is used to recompute
all eigen-values}

\item{alpha}{threshold for significance. Default = .05}

\item{col.ns}{color for the non significant
ev. Default is Green}

\item{col.sig}{color for significant eigen-values.
Default is Violet}

\item{title}{a title for the graph
default is "Explained Variance per Dimension"
# @examples  # PlotScree(ev)}
}
\description{
Plot the scree for the eigen-values
of an SVD based multivariate analysis.
Note that the function can recompute the
eigen-values when a percentage is given.
For example  ExPosition does not return all ev
       but only the requested one. but return all percentage
       so if max.ev is specified, it is used to recompute
       all eigenvalues
 If provide wit probabilities, PlotScree will
 color differently the "significant" eigenvalues
}
\author{
Hervé Abdi
}
