% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot4CATA.R
\name{PrettyBarPlot}
\alias{PrettyBarPlot}
\title{creates bar plots for a series of variables.
It is used for plotting bootstrap ratios or contributions from
principal component analysis or related methods
(e.g. correspondence analysis).}
\usage{
PrettyBarPlot(
  bootratio,
  threshold = 2,
  ylim = NULL,
  color.bar = c("lavender", "darkolivegreen3", "gray90"),
  color.letter = c("mediumpurple4", "darkolivegreen4", "gray75"),
  plotnames = TRUE,
  main = NULL,
  ylab = NULL
)
}
\arguments{
\item{bootratio}{the bootstrap ratios (BR) to be plotted
(e.g.,  obtained from \code{Boot4PTCA}).}

\item{threshold}{The critical value for significance
(default = 2, a value that matches a \eqn{p < }.05
significance level)
\code{|BR|} < threshold are plotted in gray}

\item{ylim}{a 2-element vector giving \code{min}
and \code{max} for the y-axis;
when NULL (default) set to
\code{c(min(0, min(y)), max(0,max(y)))}.}

\item{color.bar}{a 3-element vector of color names
for the bars for (respectively)
significant positive, significant negative,
and non-significant.
Default is \code{c('lavender','darkolivegreen3','gray90')}.}

\item{color.letter}{a 3-element vector of color names
for the names of the items for (respectively)
significant positive,
significant negative, and non-significant.
Default is
\code{c('mediumpurple4','darkolivegreen4','gray75')}.}

\item{plotnames}{if \code{TRUE} (default)
write the names of the items}

\item{main}{(default is \code{NULL}) a title for the graph.}

\item{ylab}{(default is \code{NULL})
a label for the y-axis (i.e., BR).}
}
\value{
A list: 1) \code{ylim}: min and max for y,
and 2) threshold.
}
\description{
\code{PrettyBarPlot}: Creates bar plots for a set of items
(e.g., observations or variables) analyzed with multivariate
methods such as PCA, CA, MCA, PLS, etc.
}
\details{
\code{PrettyBarPlot} is
used to display the bootstrap ratios or contributions plots
for CA/MCA/PCA/PLS, etc.
Significant or important items are plotted in color,
non-significant
items are plotted in gray.


\code{PrettyBarPlot2},
is a newer \code{ggplot2}-based version that can be used
\emph{in lieu} of \code{PrettyBarPlot}.
}
\seealso{
\code{\link{PrettyBarPlot2}}
}
\author{
Hervé Abdi

Hervé Abdi, Derek Beaton, and Vincent Guillemot.
}
