% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot4CATA.R
\name{PrettyBarPlotColor}
\alias{PrettyBarPlotColor}
\title{\code{PrettyBarPlotColor}: create pretty bar plots
 with defined colors for the sigificant columns}
\usage{
PrettyBarPlotColor(
  bootratio,
  threshold = 2,
  ylim = NULL,
  color4bar,
  color4ns = "gray75",
  plotnames = TRUE,
  main = NULL,
  ylab = NULL
)
}
\arguments{
\item{bootratio}{the bootstrap ratios (BR) to be plotted
(typically obtained from Boot4PTCA).}

\item{threshold}{critical.value The critical value for significance
(default = 2, which matches a $p < .05$ significance level)
|BR| < threshold are plotted in gray.}

\item{ylim}{a 2-element vector giving min and mac for y
if \code{NULL} (default) set to \code{c(-abs(BR),abs(BR))}.}

\item{color4bar}{a vector of color names (same dimension as BR)}

\item{color4ns}{(default is \code{'gray75'}) color
for the non-significant BRs.}

\item{plotnames}{if \code{TRUE} (default)
write the names of the items.}

\item{main}{(default is \code{NULL}) a title for the graph.}

\item{ylab}{(default is \code{NULL})
a label for the y axis (i.e., BR).}
}
\value{
A list: 1) \code{ylim} min and max for y,
and 2) threshold.
}
\description{
\code{PrettyBarPlotColor}: Create pretty bar plots
 with defined colors for the sigificant columns.
 NB \code{PrettyBarPlotColor} is a variation over \code{prettyBarPlot}.
}
\details{
\code{PrettyBarPlot2},
is a newer \code{ggplot2}-based plotting function
that can be used
\emph{in lieu} of \code{PrettyBarPlotColor}.
}
\seealso{
PrettyBarPlot2
}
\author{
Hervé Abdi

Vincnet Guillemeot, Hervé Abdi.
# @examples #   PrettyBarPlotColor(boot.ratio.test, color4BR)
}
