% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot4CATA.R
\name{PrettyBarPlotColor4Q}
\alias{PrettyBarPlotColor4Q}
\title{Plot the result of Cochran's Q for the columns of a CATA
table of data}
\usage{
PrettyBarPlotColor4Q(bootratio, threshold, ylim = NULL,
  color4bar = "darkorchid4", color4ns = "gray75",
  PrintSignificantOnly = TRUE, plotnames = TRUE, main = NULL,
  ylab = NULL)
}
\arguments{
\item{bootratio}{the Q values to be plotted}

\item{threshold}{(no default)
Treshold for significance, Q > threshold are plotted in gray}

\item{ylim}{the lower and upper limits for the y-axis
if NULL (default) set to c(-max(abs(y)) , max(abs(y)) )}

\item{color4bar}{the colors for the items
(default is 'darkorchid4')}

\item{color4ns}{the color for a non significant item
(default is 'gray75')}

\item{PrintSignificantOnly}{when TRUE (default)
print only the significant items}

\item{plotnames}{when TRUE (default),
plot the names in the bars}

\item{main}{a title for the plot (default is NULL)}

\item{ylab}{a label for the y axis (default is NULL)
# @examples # PrettyBarPlotColor4Q(bootratio,threshold)}
}
\description{
#' Plot the result of Cochran's Q for the columns of a CATA
table of data along with the critical values (user provided)
}
\details{
Plot the result of Cochran's Q for the columns of a CATA
table of data along with the critical values (user provided)
This version is temporary and will likely be rewriten
}
\author{
Hervé Abdi
}
