% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot4CATA.R
\name{prettyHist}
\alias{prettyHist}
\title{Plot a sampling distribution
along with a value of a criterion.}
\usage{
prettyHist(distribution, observed, show.observed = TRUE,
  observed.col = "mediumorchid4", xlim = c(-1.2, 1.2), breaks = 10,
  border = "white", distr.col = "darkseagreen", main = "", xlab = "",
  ylab = "", counts = TRUE, cutoffs = c(0.025, 0.975),
  show.cutoffs = TRUE, cutoff.col = "firebrick3", gray.distr = TRUE,
  tail.dir = "+")
}
\arguments{
\item{distribution}{the empirical distribution
(e.g., from a permutation test).}

\item{observed}{the observaed value of the criterion}

\item{show.observed}{if \code{TRUE} (default) plot the value of the criterion.}

\item{observed.col}{color for the criterion (default is "mediumorchid4").}

\item{xlim}{default is c(-1.2,1.2): value for the x-axis.
The default
works well for CA or MCA but is unlikely to work for other techniques
(such as, e.g., PCA, PLS, or Hellinger analysis).}

\item{breaks}{(default is 10) How many bins
for the histogram of the distribution.}

\item{border}{(default is "white") color for the border of the graph.}

\item{distr.col}{default is ("darkseagreen") color for the distribution.}

\item{main}{(default "") the main title.}

\item{xlab}{(default "") the x-axis label.}

\item{ylab}{(default "") the y-axis label.}

\item{counts}{if (TRUE default) shows the number of samples
if FALSE shows a probability.}

\item{cutoffs}{[default is c(0.025,0.975)],
the p-values (default for alpha = .05).}

\item{show.cutoffs}{if TRUE (default) show the critical value(s).}

\item{cutoff.col}{(default is "firebrick3") color for the critical value.}

\item{gray.distr}{if TRUE (TRUE) distribution is plotted in gray.}

\item{tail.dir}{if "+" (default) the critical value
is on the right of the
distribution, otherwise it is on the left.}
}
\value{
h the histogram of the distribution
}
\description{
Plot a sampling distribution
 (typically derived from a permutation test as
 performed, e.g., by InPosition::epCA.inference.battery).
 Plot the empirical critical value and position the
 value of the criterion.
 along with a value of a criterion.
 User is _required_ to provide a distribution
 and an observed value.
}
\author{
Derek Beaton & Hervé Abdi
}
