% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootRatio.R
\name{print.Boot4PTCA}
\alias{print.Boot4PTCA}
\title{Change the print function for Boot4PTCA}
\usage{
\method{print}{Boot4PTCA}(x, ...)
}
\arguments{
\item{x}{a list: output of Boot4PTCA}

\item{...}{everything else for the functions}
}
\description{
Change the print function for Boot4PTCA
}
\author{
Herve Abdi
}
