% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boot4Mean_etc.R
\name{print.bootGroup}
\alias{print.bootGroup}
\title{Change the print function for bootGroup}
\usage{
\method{print}{bootGroup}(x, ...)
}
\arguments{
\item{x}{a list: output of bootGroup}

\item{...}{everything else for the functions}
}
\description{
Change the print function for bootGroup
}
\author{
Herve Abdi
}
