% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootRatio.R
\name{print.bootRatios}
\alias{print.bootRatios}
\title{Change the print function for bootRatios}
\usage{
\method{print}{bootRatios}(x, ...)
}
\arguments{
\item{x}{a list: output of bootRatios}

\item{...}{everything else for the functions}
}
\description{
Change the print function for bootRatios
}
\author{
Hervé Abdi
}
