% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization4CA.R
\name{print.CArenormed}
\alias{print.CArenormed}
\title{Change the print function for CARenormalization}
\usage{
\method{print}{CArenormed}(x, ...)
}
\arguments{
\item{x}{a list: output of CARenormalization}

\item{...}{everything else for the functions}
}
\description{
Change the print function for CARenormalization
}
\author{
Herve Abdi
}
