% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFactorMaps4CA.R
\name{print.createAllMaps4CA}
\alias{print.createAllMaps4CA}
\title{Change the print function for createAllMaps4CA}
\usage{
\method{print}{createAllMaps4CA}(x, ...)
}
\arguments{
\item{x}{a list: output of createAllMaps4CA}

\item{...}{everything else for the functions}
}
\description{
Change the print function for createAllMaps4CA
}
\author{
Herve Abdi
}
