% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFactorMaps4CA.R
\name{print.createFactorMapIJ}
\alias{print.createFactorMapIJ}
\title{Change the print function for createFactorMapIJ}
\usage{
\method{print}{createFactorMapIJ}(x, ...)
}
\arguments{
\item{x}{a list: output of createFactorMapIJ}

\item{...}{everything else for the functions}
}
\description{
Change the print function for createFactorMapIJ
}
\author{
Herve Abdi
}
