% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization4CA.R
\name{print.createNormedFactors}
\alias{print.createNormedFactors}
\title{Change the print function for createNormedFactors}
\usage{
\method{print}{createNormedFactors}(x, ...)
}
\arguments{
\item{x}{a list: output of createNormedFactors}

\item{...}{everything else for the functions}
}
\description{
Change the print function for createNormedFactors
}
\author{
Herve Abdi
}
