% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSCA.R
\name{print.csca.block.part}
\alias{print.csca.block.part}
\title{Change the print function for objects of the class
\code{csca.block.part}}
\usage{
\method{print}{csca.block.part}(x, ...)
}
\arguments{
\item{x}{a list:}

\item{...}{the rest}
}
\description{
Change the print function for objects of the class
\code{csca.block.part}.
}
\author{
Herve Abdi
}
