% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSCA.R
\name{print.csca}
\alias{print.csca}
\title{Change the print function for objects of the class
\code{csca}}
\usage{
\method{print}{csca}(x, ...)
}
\arguments{
\item{x}{a list:}

\item{...}{the rest}
}
\description{
Change the print function for objects of the class
\code{csca}.
}
\author{
Herve Abdi
}
