% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSCA.R
\name{print.genCA}
\alias{print.genCA}
\title{Change the print function for objects of the class
\code{'genCA'} (e.g., from function
\code{genCA})}
\usage{
\method{print}{genCA}(x, ...)
}
\arguments{
\item{x}{a list:}

\item{...}{the rest}
}
\description{
Change the print function for objects of the class
\code{'genCA'} (e.g., from function
\code{genCA}).
}
\author{
Herve Abdi
}
