% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMap4PartialFS.R
\name{print.map4pfs}
\alias{print.map4pfs}
\title{Change the print function for the class \code{map4pfs}
(e.g., output from \code{createPartialFactorScoresMap})}
\usage{
\method{print}{map4pfs}(x, ...)
}
\arguments{
\item{x}{a list objects of the class {map4pfs}}

\item{...}{the rest}
}
\description{
Change the print function for the class \code{map4pfs}
(e.g., output from \code{createPartialFactorScoresMap}).
}
\author{
Herve Abdi
}
