% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockProjections.R
\name{print.partialProj}
\alias{print.partialProj}
\title{Change the print function for partialProj.}
\usage{
\method{print}{partialProj}(x, ...)
}
\arguments{
\item{x}{a list: output of perm4ptca}

\item{...}{everything else for the functions}
}
\description{
Change the print function for partialProj class
 objects
 (e.g.,output of partialProj4CA).
}
\author{
Herve Abdi
}
