% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perm4PTCA.R
\name{print.perm4ptca}
\alias{print.perm4ptca}
\title{Change the print function for perm4ptca class}
\usage{
\method{print}{perm4ptca}(x, ...)
}
\arguments{
\item{x}{a list: output of perm4ptca}

\item{...}{everything else for the functions}
}
\description{
Change the print function for perm4ptca class
 objects
 (output of Perm4PTCA)
}
\author{
Herve Abdi
}
