% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanZeData.R
\name{print.readCATACube.c}
\alias{print.readCATACube.c}
\title{Change the print function for class cleanCATACube.c}
\usage{
\method{print}{readCATACube.c}(x, ...)
}
\arguments{
\item{x}{a list: output of cleanCATACube}

\item{...}{the rest}
}
\description{
Change the print function for class cleanCATACube.c
(output of \code{cleanDataCATA()}
}
\author{
Herve Abdi
}
