% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCATACubeData.R
\name{print.readCATACube}
\alias{print.readCATACube}
\title{Change the print function for readCATACube}
\usage{
\method{print}{readCATACube}(x, ...)
}
\arguments{
\item{x}{a list: output of readCATACube}

\item{...}{the rest}
}
\description{
Change the print function for readCATACube
}
\author{
Herve Abdi
}
