% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockProjections.R
\name{print.rvCoeff}
\alias{print.rvCoeff}
\title{Change the print function for rvCoeff.}
\usage{
\method{print}{rvCoeff}(x, ...)
}
\arguments{
\item{x}{a list: output of \code{rvCoeff}}

\item{...}{everything else for the function}
}
\description{
Change the print function for rvCoeff class
 objects
 (e.g.,output of Rv).
}
\author{
Herve Abdi
}
