% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSymDifmat.R
\name{print.SymDif}
\alias{print.SymDif}
\title{Change the print function for object of class \code{SymDif}
(from function \code{createSymDi4PTCA})}
\usage{
\method{print}{SymDif}(x, ...)
}
\arguments{
\item{x}{a list: object of class \code{SymDif},
output of function: \code{createSymDi4PTCA}.}

\item{...}{everything else for the function}
}
\description{
Change the print function for objects of class \code{SymDif}.
 (from function \code{createSymDi4PTCA})
}
\author{
Herve Abdi
}
