% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTCA4CATA-package.R
\docType{package}
\name{PTCA4CATA-package}
\alias{PTCA4CATA-package}
\title{PTCA4CATA: Partial Triadic Analysis (PTCA) for Check All That Apply (CATA) Data}
\description{
PTCA4CATA is a collection of routines to 1) analyze Check All That Apply (CATA) data with Partial Triadic Correspondence Analysis (PTCA), 2) display the results graphically, and 3) perform the correct inferences. In addition to correspondence analysis, Hellinger analysis can also be used.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/HerveAbdi/PTCA4CATA}
  \item Report bugs at \url{http://github.com/HerveAbdi/PTCA4CATA/issues}
}

}
\author{
\strong{Maintainer}: Herve Abdi \email{herve@utdallas.edu}

Authors:
\itemize{
  \item Vincent Guillemot \email{vincent.guillemot@pasteur.fr}
  \item Ju-Chi Yu \email{ju-chi.yu@camh.ca}
}

}
\keyword{internal}
