% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTCA4CATA.R
\docType{package}
\name{PTCA4CATA}
\alias{PTCA4CATA}
\alias{PTCA4CATA-package}
\title{PTCA4CATA  package to perform Partial Triadic Correspondence Analysis
(PTCA) for Check All That Apply  (CATA) Data.}
\description{
The PTCA4CATA package provides a set of tools to analyze "Check All That Apply (CATA)" Data.
PTCA4CATA uses Partial Triadic Correspondence Analysis (PTCA) to analyze the 
pseudo-contingency table obtained from a CATA task.
}
\author{
Herve Abdi \email{herve@utdallas.edu}
}
