% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSCA.R
\name{rdiag}
\alias{rdiag}
\title{@title  right (i.e., post) Multiply a matrix by a diagonal matrix}
\usage{
rdiag(X, y)
}
\arguments{
\item{X}{an \eqn{I} by \eqn{J} matrix.}

\item{y}{a \eqn{J} element
vector (of the diagonal elements of a \eqn{J} by \eqn{J} matrix)}
}
\value{
an \eqn{I} by \eqn{J} matrix equal to
\strong{X} %*% diag(\strong{y}).
}
\description{
\code{rdiag}: right (i.e., post) Multiply
a matrix by a diagonal matrix (with only
 the diagonal elements being given).
}
\examples{
\dontrun{
 Z = rdiag(X,y)
 }
}
\author{
Herve Abdi
}
