% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCATACubeData.R
\name{read.xls.CATA}
\alias{read.xls.CATA}
\title{Read the CATA data from an Excel File and create a "Brick" of data
to be analyzed bt PTCA4CATA}
\usage{
read.xls.CATA(path2file, sheet2read, orderProducts = TRUE,
  threshold4cleaning = 0)
}
\arguments{
\item{path2file}{the name of the xls file with the data.}

\item{sheet2read}{the name of the sheet in the excel file.}

\item{orderProducts}{if TRUE (default) alphabetically
order the products.}

\item{threshold4cleaning}{the cleaning threshold:
The columns whose total
is smaller than \code{threshold4cleaning} are eliminated from
\code{CleanedContingencyTable}.
Default: \code{threshold4cleaning = 0} (i.e.,
columns with a zero sum are eliminated).
To keep even the columns with a zero sum, use
\code{threshold4cleaning = -1}}
}
\value{
A list with \code{CATA.Brick}
a "product by adjective by judge" Brick of 0/1 data
(1 if judge chose adjective for product, 0 if not);
 \code{ContingencyTable}:
A "product by adjective" contingency table;
 \code{CleanedContingencyTable}:
  A "product by adjective" cleaned contingency table
  (columns whose sum is equal or below threshold are dropped).
}
\description{
read.xls.CATA
reads the CATA data from an Excel File and creates a Brick of Data
and a contingency table. The contingency table can be analyzed
by correspondence analysis or Hellinger analysis.
The cube of data is needed to perform correct
cross-validation methos such as
permutations
tests,  bootstrap confidence intervals, and
bootstrap ratios.
With I products, J adjectives and K judges
The data in the excel file are organized as:
        row 1 column 1 = name of Judge 1
          row 1 columns  (2 to J+1) name of the adjectives
        row 2 column 1 = name of product 1.
          row 2 column (2 to J+1) 0/1 answers
            of Judge 1 to the J adjectives for product 1
        ....
        row I+1 column 1 = name of product I
           row 2 column (2 to J+1) 0/1 answers
        row I+2 column 1 = name of product 1.
          row 2 column (2 to J+1) 0/1 answers
            of Judge 2 to the J adjectives for product 1
        ....
        And so on till the last judge.
        See df ***** as data for an example
Current version uses Wickham's \code{readxl} package
and so does not need \code{rJava}.
Current version: is July 11, 2017.
}
\author{
Herve Abdi
}
