% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization4CA.R
\name{renorm4CA}
\alias{renorm4CA}
\title{Renormalize CA factor scores to
Asymmetric factor scores or true Barycentric}
\usage{
renorm4CA(X, delta, normalization = "both")
}
\arguments{
\item{X}{an I by L factor score matrix}

\item{delta}{a length L vector giving the singular value
if delta is longer than L, it is truncated and a warning is printed.}

\item{normalization}{option are '
asymmetric','barycentric','both' (default). Gives the
type of normalization.}
}
\value{
a list with F_A factor score for Asymmetric normalization
and F_B for Barycentric normalization (NB only the selected options
are returned).
}
\description{
Renormalize CA factor scores to
Asymmetric factor scores
(each dimension has norm = 1 instead of delta)
ond/or true Barycentric factor scores
(each dimension as norm = lambda instead of delta).
}
\examples{
\dontrun{
NormedFi = renorm4CA(Fi,delta)
}
# with Fi and delta from a CA program
}
\author{
Herve Abdi
}
