% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockProjections.R
\name{Rv}
\alias{Rv}
\title{\code{Rv}
Compute the RV coefficient between two matrices
with the same number of rows.}
\usage{
Rv(m1, m2, return.type = "compact")
}
\arguments{
\item{m1}{an I*J matrix}

\item{m2}{an I*K matrix}

\item{return.type}{what is in the "returned list."
\code{"compact"} (default) returns \code{Rv} and (pRv),
\code{"Rv"} returns \code{Rv}, \code{"full"} return
\code{Rv}, \code{ERv},  \code{VRv},  \code{ZRv},  \code{pRv}.}
}
\value{
a list
with, depending upon the option
chosen in \code{return.type}, some (or all) of these:
  \code{Rv} (RV coefficient),
  \code{ERv} (Expected value of Rv),
  \code{VRv} (Variance of Rv),
  \code{ZRv} (Z-score for Rv),  \code{pRv} (p-value for Rv).
}
\description{
\code{Rv}
Compute the RV coefficient between two matrices
with the same number of rows.
\code{Rv}
first computes the cross-product matrix.
With X and Y matrices, RV is equal to
<XX',YY'> /sqrt(<XX'*XX'>*<YY'*YY'>),
with <> being the scalar product.
Rv works like a squared correlation coefficient.
Depending the options chosen, \code{Rv} will return the
Rv coefficient and associated p-values.
The estimation of the p-values are made following
Kazi-Aoual et al. (CSDA 1995). See also Schlich (1996),
and Abdi (2010, 2007).
}
\references{
Abdi, H. (2007). Multiple correlation coefficient.
  In N.J. Salkind (Ed.): \emph{Encyclopedia of Measurement and Statistics.}
  Thousand Oaks (CA): Sage. pp. 648-651.

Abdi, H. (2010). Congruence: Congruence coefficient,
RV coefficient, and Mantel Coefficient.
 In N.J. Salkind, D.M., Dougherty, & B. Frey (Eds.):
  \emph{Encyclopedia of Research Design}.
  Thousand Oaks (CA): Sage. pp. 222-229.
}
\author{
Herve Abdi
}
