% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockProjections.R
\name{RV2Mat}
\alias{RV2Mat}
\title{RV2Mat
Compute the RV coefficient between two matrices
with the same number of rows.}
\usage{
RV2Mat(m1, m2)
}
\arguments{
\item{m1}{an I*J matrix}

\item{m2}{an I*K matrix}
}
\value{
rv The RV coefficient
}
\description{
\code{RV2Mat}
Compute the RV coefficient between two matrices
with the same number of rows.
\code{RV2Mat}
first computes the cross-product matrix.
With X and Y matrices, RV is equal to
<XX',YY'> /sqrt(<XX'*XX'>*<YY'*YY'>),
with <> being the scalar product.
RV works like a squared correlation coefficient.
}
\author{
Herve Abdi
}
