% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAllGraphsIntPptx.R
\name{saveGraph2pptx}
\alias{saveGraph2pptx}
\title{\code{saveGraph2pptx}:  saves all the graphics
from the Global Environment into a powerpoint file.}
\usage{
saveGraph2pptx(file2Save.pptx, title = NULL)
}
\arguments{
\item{file2Save.pptx}{the name of the powerpoint
file for saving the graphs. If this file already exists,
the old file is renamed and a warning message is printed
in the console.}

\item{title}{main title of the powerpoint file.
\code{default:} "All Graphics for Current Analysis. As of: " +
date.}
}
\value{
a list (of class \code{"savePptx"})
with  \code{listOfsavedGraphs} (the list of the graph objects
saved)  and
\code{nameOfSavingFile4pptx} (name of the files where
the graphics are saved).
}
\description{
\code{saveGraph2pptx}: a nifty function that
saves all the graphics
from the Global Environment (i.e., \code{.GlobalEnv})
into a powerpoint file. Requires packages \code{rsv}
and \code{officer}. Graphics saved are those created
with \code{recordPlot()} or \code{ggplot2}.
}
\examples{
\dontrun{
toto <- saveGraph2pptx("myFile.pptx", "Pretty Graphs of the Day")
}
}
\author{
Herve Abdi.
}
