% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot4CATA.R
\name{ShadesColor}
\alias{ShadesColor}
\title{ShadesColor Create create lighter and darker versions of a color}
\usage{
ShadesColor(AColor, jiffy = 40)
}
\arguments{
\item{AColor}{the color to use}

\item{jiffy}{A small amount to make darker or clearer
default = 40.}
}
\value{
A list:
1) TestColorDarker the darker
version of AColor
2) TestColorLighter the lighter
version of AColor
# @examples
twoColors <- ShadesColor('Red',50)
}
\description{
ShadesColor Create create lighter and darker versions of a color
 used by PrettyBarPlotColor to create faked transparent colors.
}
\author{
Hervé Abdi
}
