% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perm4PTCA.R
\name{truc4ptca}
\alias{truc4ptca}
\title{truc4ptca: a private function}
\usage{
truc4ptca(aCube, longueur = min(dim(aCube)[1:2]) - 1, permType = permType)
}
\arguments{
\item{aCube}{A cube of PTCA Data.}

\item{longueur}{the number of eigenvalues of the
data table
(default is rank of the data table -1).}

\item{permType}{the type of permutation used can be
"byRows', 'byCols', or 'byMat}
}
\description{
truc4ptca a private function:
create a permuted version of a data cube
from a PTCA analysis. Used by the function \code{perm4PTCA}.
}
\author{
Herve Abdi
}
