% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootRatio.R
\name{vec2gray}
\alias{vec2gray}
\title{transforms a vector of non-negative numbers into
gray values}
\usage{
vec2gray(levec)
}
\arguments{
\item{levec}{a vector of  non negative numbers.}
}
\value{
a vector of gray values.
}
\description{
\code{vec2gray}:
transforms a vector of non-negative numbers into
gray values.
}
\examples{
# le.grey <- vec2gray(1:10).
}
\author{
Hervé Abdi
}
