% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanZeData.R
\name{cleanDataCATA}
\alias{cleanDataCATA}
\title{cleanDataCATA
Clean a Brick of Data prior to a CATA analysis}
\usage{
cleanDataCATA(listCATA, threshold4cleaning = NULL)
}
\arguments{
\item{listCATA}{could be 1) a List as obtained,
for example, from
\code{read.xls.CATA},
or 2) a Products * Variables *  Judges array
(with only 0/1 values).}

\item{threshold4cleaning}{(default = 10% of # Judges) the minimum number
of choices for a variable to be kept.}
}
\value{
A list with \code{$CATA.Brick}
a "Product by Variables by  Judges" Brick of 0/1 data
(\emph{x(i,j,k)} = 1
if Judge \emph{k} choses Variable \emph{j}
to describe Product \emph{i}, 0 if not);
 \code{$CATA.CT}:
  A "Products by Variables" cleaned contingency table
  (columns whose sum is equal to or below
  the threshold are dropped) that gives the number of
  Judges who selected this Variable for that Product.
  \code{$nProducts} the number of Products,
  \code{$nJudges} the number of Judges, and
  \code{$nVars} the number of Variables kept.
}
\description{
\code{cleanDataCATA} cleans a \code{CATA} Brick
  of data prior to a CATA analysis.
  The CATA Brick can be a list or an array. The list can be
  obtained, for example, from \code{PTCA4CATA::read.xls.CATA}.
  If so,
  this list should include $CATA.Brick.
  If the data is an array, it should be
  a Products * Variables * Judges array.
  The level of "almost empty" is defined by the parameter
  \code{threshold4cleaning}
  (set by default to 10% of the number of judges).
}
\examples{
\dontrun{
data('beersBeTACATA')
rawCubeOfData     <- beersBeTACATA$rawBrick.CATA
cleanedListOfData <- cleanDataCATA(rawCubeOfData, threshold4cleaning =  5)
}
}
\author{
Herve Abdi
}
