% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFactorMaps4CA.R
\name{createAllMaps4CA}
\alias{createAllMaps4CA}
\title{Creates all the partial ggplot2 maps for CA
with all standard variants of normalization for factor scores.}
\usage{
createAllMaps4CA(allNormedFactors, axis1 = 1, axis2 = 2,
  title = NULL, col.points.i = "blueviolet", alpha.points.i = 0.5,
  pch.i = 19, cex.i = 2.5, segment.size.i = 0,
  col.labels.i = "darkorchid4", alpha.labels.i = 1, text.cex.i = 4,
  font.face.i = "bold", font.family.i = "sans", force.i = 1,
  col.points.j = "darkolivegreen4", alpha.points.j = 0.5, pch.j = 18,
  cex.j = 2.5, segment.size.j = 0, col.labels.j = "darkolivegreen",
  alpha.labels.j = 1, text.cex.j = 4, font.face.j = "bold",
  font.family.j = "sans", force.j = 1, col.axes = "darkorchid",
  alpha.axes = 0.2, width.axes = 1.1,
  col.background = adjustcolor("lavender", alpha.f = 0.2), ...)
}
\arguments{
\item{allNormedFactors, }{A list with all the
factor scores normed
typically from \code{createAllNormedFactors}.}

\item{axis1}{the column of X used for the horizontal axis
of the plot. Default =  1.}

\item{axis2}{the column of X used for the vertical axis
of the plot. Default = 2.}

\item{title}{A main title (default is \code{NULL}).}

\item{col.points.i}{the color of the points/dots
for the \eqn{I}-set.
Can be one color or a vector of colors. If a vector, it needs
to have exactly the number of items to be plotted.
Default = \code{'blueviolet'}.}

\item{alpha.points.i}{(default = .5), the alpha
(transparency) for the points, should be
between 1 (opaque) and 0
(completely transparent).}

\item{pch.i}{the character for the points for the \eqn{I}-set.
Default is 19 (Circles).}

\item{cex.i}{size of the dots for the \eqn{I}-set.
Default = 2.5}

\item{segment.size.i}{= 0, # size of segment}

\item{col.labels.i}{the color of the labels for
the \eqn{I}-set.
Can be one color or a vector of colors. If a vector, it needs
to have exactly the number of items to be plotted.
Default =  \code{'darkorchid'}.}

\item{alpha.labels.i}{(default = 1), the alpha
(transparency) for the labels, should be
between 1 (opaque) and 0
(completely transparent).}

\item{text.cex.i}{= 4,  font size for labels for the I-set.}

\item{font.face.i}{(Default = \code{'bold'})
font for labels for the \eqn{I}-set.}

\item{font.family.i}{(Default = \code{'sans'})
font family for the \eqn{I}-set.}

\item{force.i}{(Default = 1)
How much \code{ggrepel} repels the labels
for the  \eqn{I}-set.}

\item{col.points.j}{the color of the points/dots for the \eqn{J}-set.
Can be one color or a vector of colors. If a vector, it needs
to have exactly the number of items to be plotted.
Default = \code{'darkolivegreen4'}.}

\item{alpha.points.j}{(Default = .5), the alpha
(transparency) for the \eqn{J}-points, should be
between 1 (opaque) and 0
(completely transparent).}

\item{pch.j}{the character for the points for the \eqn{J}-set.
Default is 18 (Diamonds).}

\item{cex.j}{size of the dots for the \eqn{J}-set.
Default = 2.5}

\item{segment.size.j}{= 0, # size of segment for \eqn{J}-set}

\item{col.labels.j}{the color of the labels for
the \eqn{J}-set.
Can be one color or a vector of colors. If a vector, it needs
to have exactly the number of items to be plotted.
Default =  \code{'darkolivegreen'}.}

\item{alpha.labels.j}{(default = 1), the alpha
(transparency) for the \eqn{J}-labels, should be
between 1 (opaque) and 0
(completely transparent).}

\item{text.cex.j}{(Default = 4)
font size for labels for the \eqn{J}-set.}

\item{font.face.j}{(Default = \code{'bold'}
font for the labels for the \eqn{J}-set.}

\item{font.family.j}{(Default = \code{'sans'})
font family for the \eqn{J}-set.}

\item{force.j}{(Default = 1). How much \code{ggrepel} repels
the labels for the \eqn{J}-set.}

\item{col.axes}{(Default = \code{'darkorchid'})
color for the axes.}

\item{alpha.axes}{(Default = .2) transparency factor for the
color of the axes.}

\item{width.axes}{(Default = 1.1) the width of the axes}

\item{col.background}{[Default =  \code{adjustcolor('lavender', alpha.f = .2)}],
The color of the background.}

\item{...}{eveythings else for the functions.}
}
\value{
a list with
\code{baseMap_S:} The Background (Symmetric)
\code{I_labels_S}  The labels I-set (Symmetric).
\code{I_points_S}  The dots I-set (Symmetric).
\code{J_labels_S}  The labels J-set (Symmetric).
 \code{J_points_S} The dots J-set(Symmetric).
 \code{baseMap_A} The Background (Asymmetric)
  \code{I_labels_A}  The labels I-set (Asymmetric).
\code{I_points_A} The dots I-set (Asymmetric).
  \code{J_labels_A}  The labels J-set (Asymmetric).
  \code{J_points_A}  The dots J-set (Asymmetric).
 \code{baseMap_B} The Background (Barycentric).
  \code{I_labels_B} The labels I-set (Barycentric).
  \code{I_points_B}  The dots I-set (Barycentric).
  \code{J_labels_B} The labels J-set (Barycentric).
 \code{J_points_B}  The dots J-set (Barycentric).
}
\description{
\code{createAllMaps4CA}: uses
 \code{ggplot2} to
 1) create all the partial maps for CA
with the different types of normalizations: Asymmetric,
Symmetric, True-Barycentric, Biplots, and \code{SPSS}
 pseudo Biplots.
2)
  to creates \code{ggplot2} factorial maps for dots and labels;
3) to create the base plot maps for CA type graphs, and 4)
to creates maps for the \eqn{I} (rows) and the
\eqn{J} (column) -sets.
}
\details{
The final maps are built
by combining the elementary maps.
For example, a map with the \eqn{J}-set being asymmetric
and the \eqn{I}-set being Symmetric would be made
by creating the \code{Ja_Is} map as
\code{Ja_Is <- baseMap_A + J_labels_A + I_labels_S}.

Note than, in general, the \code{Asymmetric} is larger than the
\code{Symmetric} map which is in turn larger than the
\code{Barycentric} map.
}
\section{Important_Note}{

When creating multiple layers graphs,
because of the way \code{ggplot2} create graphs,
all
the matrices/dataframe should all the have
the  same column names
[e.g., \code{colnames()}
equal to c("Dimension 1", "Dimension 2")].
When this is not the case, some strange and cryptic
errors may be produced
(e.g., "cannot find Dimension").
}

\seealso{
createFactorMaps createbaseMap
}
\author{
Hervé Abdi
}
