% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFactorMaps4CA.R
\name{createBaseMap}
\alias{createBaseMap}
\title{create a base map for CA type graphs with ggplot2}
\usage{
createBaseMap(
  data,
  constraints = NULL,
  col.axes = "darkorchid",
  alpha.axes = 0.2,
  width.axes = 1.1,
  col.background = adjustcolor("lavender", alpha.f = 0.2),
  title = NULL
)
}
\arguments{
\item{data}{the factor scores to plot}

\item{constraints}{a list with minx miny maxx maxy
typically obtained from  \code{prettyGraphs::minmaxHelper()}.
If \code{NULL}
(default) it  is computed with the function
 \code{prettyGraphs::minmaxHelper()}.}

\item{col.axes}{color for the axes, default is
\code{'darkorchid'}.}

\item{alpha.axes}{alpha parametere (transparency)
for the axes, default is .2.}

\item{width.axes}{the width of the axes, default is 1.1.}

\item{col.background}{the color theme of the background,
default is \code{adjustcolor('lavender', alpha.f = .2)}.}

\item{title}{a main title, default is \code{NULL} (no title)}
}
\value{
a basemap
}
\description{
\code{createBaseMap}: Create a ggplot2 basemap for CA type graphs.
The final maps are created by using overlays.
A map CA is created with first the baseMap and
then adding a text/dot map.
}
\section{Important_Note}{
 When creating multiple layers graphs,
because of the way \code{ggplot2} create graphs, all the
the matrices/dataframe should all the have the same column names
[e.g., \code{colnames()} equal to
\code{c("Dimension 1", "Dimension 2")}].
When it is not the case, some strange and cryptic
error may be produced
(e.g., \code{"cannot find Dimension"}).
}

\examples{
\dontrun{
aBaseMap <- createBaseMap(Fi)
# with Fi being a map of factor scores
}
}
\seealso{
\code{\link{createFactorMapIJ}}
}
\author{
Hervé Abdi
}
