% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCmat.R
\name{createCmat4PTCA}
\alias{createCmat4PTCA}
\title{Create a matrix of cross-product or RV coefficients
for the 3rd dimension of a brick of non-negative numbers
(i.e., a brick of data for a CATA test).}
\usage{
createCmat4PTCA(dataCube, normalization = "cos")
}
\arguments{
\item{dataCube}{an I*J*K brick of non-negative numbers
(i.e., a brick of data for a CATA test)}

\item{normalization}{Type of normalization
can be 'cos' (Default) or 'Rv' (for the Rv coefficient)}
}
\value{
An I*I cosine or Rv Matrix
}
\description{
\code{createCmat4PTCA}
Creates a matrix of cross-product or RV coefficients
for the 3rd dimension of an I*J*K brick of non-negative numbers
(i.e., a brick of data for a CATA test).
The coefficients
are computed from the I row-profiles (observations) and stored
in a I*I semi-positive definite matrix that can be analyzed by
an eigen-decomposition to provide a STATIS-like RV/cosine-map.
}
\details{
Each of the K slices of the I*J*K brick
of data is first transformed into an I*I S matrix using the
function \code{PTCA4CATA::CA.SfromX}
 (Empty columns are eliminated before computing the matrix S).
 This creates an I*I*K brick of S matrices wichi then used to
 compute the I*I cosine/Rv matrix that measures the similarity
 between all slice of \code{dataCube}.
 Note that matrice can used in a STATIS apporach to re-weight the
 slices of the \code{dataCube}.
 Note that the rows of each silice are supposed to have
 at least one non-zero entry.
 Slices with zeros rows are eliminated and a warning
 message is issued
}
\examples{
\dontrun{
if(interactive()){
 # with aCubeOfCATAData being an I*J*K array
 Cmat <- createCmat4PTCA(aCubeOfCATAData)
 }
}
}
\author{
Herve Abdi
}
