% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot4CATA.R
\name{createLabel}
\alias{createLabel}
\title{Creates\eqn{x} and \eqn{y} labels for ggplots2 scatterplots}
\usage{
createLabel(resCA, zeAxis, axisName = "Dimension")
}
\arguments{
\item{resCA}{the results of \code{epCA} or \code{epPCA}.}

\item{zeAxis}{the number of the axis (no default)}

\item{axisName}{the name for the axes (default = \code{'Dimension'})}
}
\description{
\code{createLabel} creates \eqn{x} and \eqn{y}
labels for \code{ggplots2} scatterplots
(e.g., correspondence analysis or principal component analysis).
}
\author{
Herve Abdi
}
