% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot4CATA.R
\name{PrettyBarPlot}
\alias{PrettyBarPlot}
\title{\code{PrettyBarPlot}: creates bar plots for a series of variables.
It is use for plotting bootstrap ratios or contributions from
principal component analysis or related methods
(e.g. correspondence analysis).}
\usage{
PrettyBarPlot(bootratio, threshold = 2, ylim = NULL,
  color.bar = c("lavender", "darkolivegreen3", "gray90"),
  color.letter = c("mediumpurple4", "darkolivegreen4", "gray75"),
  plotnames = TRUE, main = NULL, ylab = NULL)
}
\arguments{
\item{bootratio}{the bootstrap ratios (BR) to be plotted
(e.g.,  obtained from \code{Boot4PTCA}).}

\item{threshold}{critical.value The critical value for significance
(default = 2, which matches a $p < .05$ significance level)
|BR| < threshold are plotted in gray}

\item{ylim}{a 2-element vector giving min and max for the y-axis;
when NULL (default) set to c(-max(abs(y)),max(abs(y)))}

\item{color.bar}{a 3-element vector of color names
for the bars for (respectively)
significant positive, significant negative, and non-significant.
Default is c('lavender','darkolivegreen3','gray90')}

\item{color.letter}{a 3-element vector of color names
for the names of the items for (respectively)
significant positive, significant negative, and non-significant.
Default is c("mediumpurple4",'darkolivegreen4','gray75')}

\item{plotnames}{if TRUE (default) write the names of the items}

\item{main}{(default is NULL) a title for the graph}

\item{ylab}{(default is NULL) a label for the y axis (i.e., BR)}
}
\value{
A list: 1) ylim min and max for y, 2) threshold
}
\description{
\code{PrettyBarPlot}: creates bar plots for a set of items
(e.g., observations or variables) analyzed with multivariate
methods such as PCA, CA, MCA, PLS etc.
 \code{PrettyBarPlot} is
used to display the bootstrap ratios or contributions plots
for CA/MCA/PCA/PLS.
Significant or important items are plotted in color, non-significant
items are plotted in gray.
}
\author{
Hervé Abdi
}
