% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrettyBarPlot2.R
\name{PrettyBarPlot2}
\alias{PrettyBarPlot2}
\title{a \code{ggplot2}-based function to plot bar plots for a series
of variables:
Used for plotting bootstrap ratios or contributions in PCA, PLS, etc.}
\usage{
PrettyBarPlot2(bootratio, font.size = NULL, threshold = 3,
  ylim = NULL, color.bar = c("lavender", "darkolivegreen3", "gray90"),
  color.bord = c("mediumpurple4", "darkolivegreen4", "gray75"),
  color.letter = c("mediumpurple4", "darkolivegreen4", "gray75"),
  color4ns = "gray75", color4bar = NULL, plotnames = TRUE,
  main = NULL, ylab = NULL, sortValues = FALSE, signifOnly = FALSE,
  angle.text = 90, horizontal = TRUE)
}
\arguments{
\item{bootratio}{the bootstrap ratios (BR) or contributions
or similar statistics to be plotted.
(e.g., obtained from Boot4PTCA).}

\item{font.size}{PARAM_DESCRIPTION, Default: NULL}

\item{threshold}{(Default: 2)
The critical value for significance
 which matches a \eqn{p} < .05
significance level. The |BR| < threshold are plotted in gray,
the other BR are plotted in colors.}

\item{ylim}{(Default: NULL)
a 2-element vector giving min and max for the y-axis;
when NULL (default) set to c(-max(abs(y)), max(abs(y)))}

\item{color.bar}{a 3-element vector
of color names for the bars for (respectively)
significant positive,
significant negative, and non-significant.
Default is c('lavender','darkolivegreen3','gray90').}

\item{color.bord}{The color for the boder of the bars,
Default: c("mediumpurple4", "darkolivegreen4", "gray75").}

\item{color.letter}{a 3-element vector of color names for the names of
the items for (respectively)
significant positive, significant negative, and non-significant.
Default is c("mediumpurple4",'darkolivegreen4','gray75')}

\item{color4ns}{(Default: \code{'gray75'})
color for the non-significant \code{bootratio}.}

\item{color4bar}{(Default: NULL) a vector of color names
(same dimension as \code{bootratio})}

\item{plotnames}{if TRUE (default) write the names of the items}

\item{main}{(default is \code{NULL}) a title for the graph}

\item{ylab}{(default is \code{NULL}) a label for the \eqn{y} axis (i.e., BR).
If \code{NULL} \code{ylab = 'Bootstrap ratios'}.}

\item{sortValues}{(Default: \code{FALSE}) when \code{TRUE}
sort the values to plot.}

\item{signifOnly}{(Default: \code{FALSE}) when \code{TRUE}
plot only the significant values.}

\item{angle.text}{(Default: \code{90}, i.e., so slant)
the "slanting factor" for the text.}

\item{horizontal}{(Default: \code{TRUE}) when \code{TRUE},
the plot is horizontal, when \code{FALSE} the plot is vertical.}
}
\value{
A \code{ggplot2} object containg the graph
(i.e., to be plotted with \code{print}).
}
\description{
\code{PrettyBarPlot2}:
a \code{ggplot2}-based version of \code{PrettybarPlot} and
\code{PrettyBarPlotColor}.
\code{PrettyBarPlot2}
creates bar plots for a set of items
(e.g., observations or variables) analyzed with multivariate methods
such as PCA, CA, MCA, PLS etc.
PrettyBarPlot is used to display the bootstrap ratios
or contributions plots for CA/MCA/PCA/PLS.
Significant or important items are plotted
in color, non-significant items are plotted in gray.
}
\details{
\code{PrettyBarPlot2} intergrates \code{PrettyBarPlot}
and \code{PrettyBarPlotColors}.
}
\examples{
toto <- 8*runif(7)
names(toto) <- paste0('V', 1:7)
PrettyBarPlot2(toto)
}
\seealso{
\code{\link{PrettyBarPlot}}
\code{\link{PrettyBarPlotColor}}
}
\author{
Vincent Guillemot & Hervé Abdi
}
