% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot4CATA.R
\name{PrettyBarPlotColor4Q}
\alias{PrettyBarPlotColor4Q}
\title{\code{PrettyBarPlotColor4Q}
Plot the result of Cochran's \eqn{Q} for the columns of a CATA
table of data.}
\usage{
PrettyBarPlotColor4Q(
  bootratio,
  threshold,
  ylim = NULL,
  color4bar = "darkorchid4",
  color4ns = "gray75",
  PrintSignificantOnly = TRUE,
  plotnames = TRUE,
  main = NULL,
  ylab = NULL
)
}
\arguments{
\item{bootratio}{the \eqn{Q} values to be plotted}

\item{threshold}{(no default)
Treshold for significance, Q > threshold are plotted in gray}

\item{ylim}{the lower and upper limits for the y-axis
if NULL (default) set to c(-max(abs(y)) , max(abs(y)) )}

\item{color4bar}{the colors for the items
(default is 'darkorchid4')}

\item{color4ns}{the color for a non significant item
(default is 'gray75')}

\item{PrintSignificantOnly}{when \code{TRUE} (default)
print only the significant items}

\item{plotnames}{when \code{TRUE} (default),
plot the names in the bars}

\item{main}{a title for the plot (default is \code{NULL})}

\item{ylab}{a label for the \eqn{y} axis (default is \code{NULL})
# @examples # PrettyBarPlotColor4Q(bootratio,threshold)}
}
\description{
\code{PrettyBarPlotColor4Q}:
Plot the result of Cochran's \eqn{Q} for the columns
of a \code{CATA}
table of data along with the critical value (user provided)
NB: This version is temporary and will likely be rewriten
(see \code{PrettybarPlot2} for a \code{ggplot2}-based
equivalent function).
}
\seealso{
PrettyBarPlot2
}
\author{
Hervé Abdi
}
