% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTCA4CATA.R
\docType{package}
\name{PTCA4CATA}
\alias{PTCA4CATA}
\title{PTCA4CATA  package to perform Partial Triadic Correspondence Analysis
(PTCA) for Check All That Apply (CATA) Data}
\description{
The PTCA4CATA package provides a set of tools to 1) analyze
"Check All That Apply (CATA)" Data;
   2) display the results graphically;
and 3)
perform the correct inferences.
In addition to correspondence analysis, Hellinger analysis
can also be used.
PTCA4CATA uses Partial Triadic Correspondence Analysis (PTCA)
to analyze the
pseudo-contingency table obtained from a CATA task.
}
\author{
Herve Abdi \email{herve@utdallas.edu}
}
