% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function4Q.R
\name{Q4CATA.Cube}
\alias{Q4CATA.Cube}
\title{Compute Cochran's Q
for a Cube of CATA Data}
\usage{
Q4CATA.Cube(ZeCube)
}
\arguments{
\item{ZeCube}{A cube of 0/1 CATA data}
}
\value{
a 4 * J matrix.
Row 1 gives the value of chi2
Row 2 gives the uncorrected p-value
Row 3 gives the  Sidak corrected (for multiple
comparisons)  p-value
Row 4 gives the  Bonferroni corrected (for multiple
comparisons)  p-value
}
\description{
Compute Cochran's Q
for a Cube of CATA Data
(as created, e.g., by \code{readCATAfromXL})
I: Rows   are Products
J: Columns are Variables
K: Third Dimension is Assessors.
x_{i,j,k} = 1 means
Assessor k chosed Variable j for Product i
NB. uses function \code{Q4CATA.Slice}
}
\author{
Herve Abdi
}
