% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function4Q.R
\name{Q4CATA.Slice}
\alias{Q4CATA.Slice}
\title{Cochran's Q test for a slice of a cube}
\usage{
Q4CATA.Slice(Data4Q)
}
\arguments{
\item{Data4Q}{a slice of CATA cube}
}
\value{
a vector with the value of \code{chi2}
 and \code{pvalue}. The value for the chi2
}
\description{
Cochran's Q test for a slice of a cube.
 Used as a subfunction of Q4CATA.Cube
 Needs the \code{coin} library.
 The data are stored in a matrix
that represents the choices of
J participants (col) for I products (row)
about one variable
with
    rows as products
    columns as assessors
x(i,j) = 1 if p(i) was chosen by a(j), 0 if not.
}
\author{
Herve Abdi
}
