% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockProjections.R
\name{RV4Brick}
\alias{RV4Brick}
\title{Compute the \eqn{RV}
coefficient between the slices of a data array.}
\usage{
RV4Brick(aBrickOfData)
}
\arguments{
\item{aBrickOfData}{a brick of data
\eqn{I*J_k*K}. The \eqn{RV} coefficient
is computed between the \eqn{K}-slices of the brick}
}
\value{
a list with a matrix of the \eqn{RV} coefficients
(\code{$RV}), and
a matrix of \eqn{p}-values ((\code{$pRV})).
}
\description{
\code{RV4Brick} computes the RV coefficient
between the slices of a data array,
such as block projection
for CA in PTCA4CATA.
}
\author{
Hervé Abdi
}
