% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot4CATA.R
\name{ShadesColor}
\alias{ShadesColor}
\title{Create lighter and darker versions of a color.}
\usage{
ShadesColor(aColor, jiffy = 40)
}
\arguments{
\item{aColor}{the color to use}

\item{jiffy}{A small amount to make darker or clearer
default = 40.}
}
\value{
A list:
1) \code{testColorDarker}: the darker
version of \code{aColor};
2) \code{testColorLighter} the lighter
version of \code{aColor}.
}
\description{
\code{ShadesColor}: Create lighter and darker
versions of a color
 used by \code{PrettyBarPlotColor}
 to create faked transparent colors.
}
\examples{
twoColors <- ShadesColor('Red',50)
}
\author{
Hervé Abdi
}
